% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates_auxiliary.R
\name{calc_apply_time_summary}
\alias{calc_apply_time_summary}
\title{Apply default/native or explicit temporal summarization}
\usage{
calc_apply_time_summary(
  covar,
  .by_time = NULL,
  fun_summary = "mean",
  locs_id = "site_id",
  time_col = "time",
  group_cols_extra = NULL
)
}
\arguments{
\item{covar}{data.frame.}

\item{.by_time}{NULL or character(1).}

\item{fun_summary}{character(1) or function.}

\item{locs_id}{character(1).}

\item{time_col}{character(1).}

\item{group_cols_extra}{character or NULL.}
}
\value{
data.frame
}
\description{
Apply default/native or explicit temporal summarization
}
\keyword{auxiliary}
\keyword{internal}
