% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates_auxiliary.R
\name{calc_check_time}
\alias{calc_check_time}
\title{Check time values}
\usage{
calc_check_time(covar, POSIXt = TRUE)
}
\arguments{
\item{covar}{data.frame(1). Calculated covariates \code{data.frame}.}

\item{POSIXt}{logical(1). Should the time values in \code{covar} be of class
\code{POSIXt}? If \code{FALSE}, the time values will be checked for integer class
(year and year-month).}
}
\description{
Check the time values within calculated covariates \code{data.frame}
}
\keyword{internal}
