% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R, R/process.R
\name{calc_covariates}
\alias{calc_covariates}
\title{Calculate covariates}
\usage{
calc_covariates(
  covariate = c("modis", "koppen-geiger", "koeppen-geiger", "koppen", "koeppen", "geos",
    "dummies", "gmted", "roads", "sedac_groads", "nlcd", "tri", "ncep", "nei",
    "ecoregions", "ecoregion"),
  path,
  ...
)

calc_covariates(
  covariate = c("modis", "koppen-geiger", "koeppen-geiger", "koppen", "koeppen", "geos",
    "dummies", "gmted", "roads", "sedac_groads", "nlcd", "tri", "ncep", "nei",
    "ecoregions", "ecoregion"),
  path,
  ...
)
}
\arguments{
\item{covariate}{character(1). Covariate type.}

\item{path}{character(1). Directory or file path to raw data
depending on \code{covariate} value.}

\item{...}{Arguments passed to each raw data processing function.
function.}

\item{locs}{sf/SpatVector. Unique locations. Should include
a unique identifier field named \code{locs_id}}

\item{from}{character. Single or multiple from strings.}

\item{locs_id}{character(1). Name of unique identifier.
Default is \code{"site_id"}.}
}
\value{
Calculated covariates. Mainly data.frame object.

\code{SpatVector} or \code{SpatRaster} depending on covariate type.
}
\description{
Calculate covariates

Process covariates
}
\seealso{
\itemize{
\item \code{calc_modis}: \code{"modis"}, \code{"MODIS"}
\item \code{calc_koppen_geiger}: \code{"koppen-geiger"}, \code{"koeppen-geiger"}, \code{"koppen"},
\item \code{calc_ecoregion}: \code{"ecoregion"}, \code{"ecoregions"}
\item \code{calc_temporal_dummies}: \code{"dummies"}
}

\itemize{
\item \code{process_modis}: \code{"modis"}, \code{"MODIS"}
\item \code{process_koppen_geiger}: \code{"koppen-geiger"}, \code{"koeppen-geiger"}, \code{"koppen"},
\item \code{process_ecoregion}: \code{"ecoregion"}, \code{"ecoregions"}
}
}
\author{
Insang Song
}
