% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_ecoregion}
\alias{calc_ecoregion}
\title{Calculate ecoregions covariates}
\usage{
calc_ecoregion(from = NULL, locs, locs_id = "site_id", geom = FALSE, ...)
}
\arguments{
\item{from}{SpatVector(1). Output of \code{\link{process_ecoregion}}.}

\item{locs}{sf/SpatVector. Unique locs. Should include
a unique identifier field named \code{locs_id}}

\item{locs_id}{character(1). Name of unique identifier.}

\item{geom}{logical(1). Should the geometry of \code{locs} be returned in the
\code{data.frame}? Default is \code{FALSE}. If \code{geom = TRUE} and \code{locs} contain
polygon geometries, the \verb{$geometry} column in the returned data frame may
make the \code{data.frame} difficult to read due to long geometry strings. The
coordinate reference system of the \verb{$geometry} is the coordinate
reference system of \code{from}.}

\item{...}{Placeholders.}
}
\value{
a data.frame object with dummy variables and attributes of:
\itemize{
\item \code{attr(., "ecoregion2_code")}: Ecoregion lv.2 code and key
\item \code{attr(., "ecoregion3_code")}: Ecoregion lv.3 code and key
}
}
\description{
Extract ecoregions covariates (U.S. EPA Ecoregions Level 2/3) at point
locations. Returns a \code{data.frame} object containing \code{locs_id} and
binary (0 = point not in ecoregion; 1 = point in ecoregion) variables for
each ecoregion.
}
\seealso{
\code{\link{process_ecoregion}}
}
\author{
Insang Song
}
