% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_example}
\alias{calc_example}
\title{Example function to increase length}
\usage{
calc_example(x)
}
\arguments{
\item{x}{character.}
}
\description{
Example function to increase length
}
\keyword{auxiliary}
