% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates_auxiliary.R
\name{calc_extents_overlap}
\alias{calc_extents_overlap}
\title{Validate extents overlap}
\usage{
calc_extents_overlap(x, y)
}
\arguments{
\item{x}{SpatRaster(1)}

\item{y}{SpatRaster(1)}
}
\value{
logical(1)
}
\description{
Validate extents overlap
}
\keyword{auxiliary}
\keyword{internal}
