% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_hms}
\alias{calc_hms}
\title{Calculate NOAA HMS Wildfire Smoke covariates}
\usage{
calc_hms(from, locs, locs_id = NULL, radius = 0)
}
\arguments{
\item{from}{SpatVector(1). Cleaned SpatVector object that has been returned
from \code{process_hms} containing wildfire smoke plume coverage data.}

\item{locs}{data.frame, characater to file path, SpatVector, or sf object.}

\item{locs_id}{character(1). Column within \code{locations} CSV file
containing identifier for each unique coordinate location.}

\item{radius}{integer(1). Circular buffer distance around site locations.
(Default = 0).}
}
\value{
a data.frame object;
}
\description{
Extract wildfire smoke plume coverage data from NOAA Hazard Mapping Fire and
Smoke Product at point locations using SpatVector object from \code{process_hms}.
Function returns a data frame containing wildfire smoke plume binary values
(0 = smoke absent; 1 = smoke present) at user-defined sites. Unique columns
reflect smoke density and circular buffer.
}
\author{
Mitchell Manware
}
