% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates_auxiliary.R
\name{calc_linear_unit_metadata}
\alias{calc_linear_unit_metadata}
\title{Describe linear units used for covariate calculations}
\usage{
calc_linear_unit_metadata(crs)
}
\arguments{
\item{crs}{Coordinate reference system accepted by \code{terra::crs()}.}
}
\value{
A named list describing CRS and buffer units.
}
\description{
Describe linear units used for covariate calculations
}
\keyword{auxiliary}
\keyword{internal}
