% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates_auxiliary.R
\name{calc_message}
\alias{calc_message}
\title{Send progress messages}
\usage{
calc_message(dataset, variable, time, time_type, level)
}
\arguments{
\item{dataset}{character(1). Data source.}

\item{variable}{placeholder}

\item{time}{placeholder}

\item{time_type}{placeholder}

\item{level}{placeholder}
}
\value{
NULL; provides progress messages to R console.
}
\description{
Send messages updating covariate extraction progress.
}
\keyword{auxiliary}
\keyword{internal}
