% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_modis}
\alias{calc_modis}
\title{Calculate MODIS product covariates in multiple CPU threads}
\usage{
calc_modis(
  path,
  product = c("MOD11A1", "MOD13A2", "MOD06_L2", "VNP46A2", "MOD09GA", "MCD19A2"),
  sites,
  id_col = "site_id",
  name_covariates,
  radius = c(0L, 1000L, 10000L, 50000L),
  subdataset = NULL,
  fun_summary = "mean",
  nthreads = floor(length(parallelly::availableWorkers())/2),
  package_list_add = NULL,
  export_list_add = NULL
)
}
\arguments{
\item{path}{character. List of HDF files.}

\item{product}{character(1). MODIS product. Should be one of
\code{c("MOD11A1", "MOD13A2", "MOD06_L2", "VNP46A2", "MOD09GA", "MCD19A2")}}

\item{sites}{sf object. Unique sites where covariates
will be calculated.}

\item{id_col}{character(1). Site identifier. Default is \code{"site_id"}}

\item{name_covariates}{character. Name header of covariates.
e.g., \code{"MOD_NDVIF_0_"}.
The calculated covariate names will have a form of
'{name_covariates}{zero-padded buffer radius in meters}',
e.g., 'MOD_NDVIF_0_50000' where 50 km radius circular buffer
was used to calculate mean NDVI value.}

\item{radius}{numeric. Radii to calculate covariates.
Default is \code{c(0, 1000, 10000, 50000)}.}

\item{subdataset}{Index or search pattern of subdataset.
Pattern is not accepted when product is \code{"VNP46A2"} or
\code{"MOD06_L2"}}

\item{fun_summary}{character or function. Function to summarize
extracted raster values.}

\item{nthreads}{integer(1). Number of threads to be used
to calculate covariates.}

\item{package_list_add}{character. A vector with package names to load
these in each thread. Note that \code{sf}, \code{terra},
\code{exactextractr}, \code{doParallel}, \code{parallelly},
and \code{dplyr}
are the default packages to be loaded.}

\item{export_list_add}{character. A vector with object names to export
to each thread. It should be minimized to spare memory.}
}
\description{
calc_modis essentially runs \code{modis_worker} function
in each thread (subprocess). Based on daily resolution, each day's workload
will be distributed to each thread. With \code{product} argument,
the files are processed by a customized function where the unique structure
and/or characteristics of the products are considered. \code{nthreads}
argument should be carefully selected in consideration of the machine's
CPU and memory capacities as products have their own memory pressure.
Overall, this function and dependent routines assume that the file system
can handle concurrent access to the (network) disk by multiple processes.
File system characteristics, package versions, and hardware settings/
specification can affect the processing efficiency.

sites should be sf object as it is exportable to
parallel workers.
}
\note{
See details for setting parallelization
\code{\link[foreach]{foreach}},
\code{\link[parallelly]{makeClusterPSOCK}},
\code{\link[parallelly]{availableCores}},
\code{\link[doParallel]{registerDoParallel}}
}
