% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_nlcd_ratio}
\alias{calc_nlcd_ratio}
\title{Compute land cover classes ratio in circle buffers around points}
\usage{
calc_nlcd_ratio(
  locs,
  from,
  locs_id = "site_id",
  radius = 1000,
  variables = NULL
)
}
\arguments{
\item{locs}{terra::SpatVector of points geometry}

\item{from}{SpatRaster of NLCD}

\item{locs_id}{character(1). Unique identifier of locations}

\item{radius}{numeric (non-negative) giving the
radius of buffer around points}

\item{variables}{character. Variable names to extract. Ignored.}

\item{year}{numeric giving the year of NLCD data used}
}
\description{
Compute land cover classes ratio in circle buffers around points
}
