% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates_auxiliary.R
\name{calc_prepare_exact_geoms}
\alias{calc_prepare_exact_geoms}
\title{Convert point extractions to tiny polygons for exact extraction}
\usage{
calc_prepare_exact_geoms(locs_vector, radius)
}
\arguments{
\item{locs_vector}{SpatVector(1)}

\item{radius}{numeric(1)}
}
\value{
sf object for exactextractr
}
\description{
Convert point extractions to tiny polygons for exact extraction
}
\keyword{auxiliary}
\keyword{internal}
