% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates_auxiliary.R
\name{calc_prepare_weights}
\alias{calc_prepare_weights}
\title{Prepare optional weighting raster}
\usage{
calc_prepare_weights(from, weights = NULL)
}
\arguments{
\item{from}{SpatRaster(1). Template raster.}

\item{weights}{NULL, SpatRaster, SpatVector/sf polygon, or file path.}
}
\value{
NULL or single-layer SpatRaster aligned to \code{from}.
}
\description{
Prepare optional weighting raster
}
\keyword{auxiliary}
\keyword{internal}
