% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_sedc}
\alias{calc_sedc}
\title{Calculate Sum of Exponentially Decaying Contributions (SEDC) covariates}
\usage{
calc_sedc(
  locs = NULL,
  from = NULL,
  locs_id = NULL,
  sedc_bandwidth = NULL,
  target_fields = NULL
)
}
\arguments{
\item{locs}{\code{SpatVector} object. Locations where
the sum of SEDCs are calculated.}

\item{from}{\code{SpatVector} object. Locations where each SEDC is calculated.}

\item{locs_id}{character(1). Name of the unique id field in \code{point_to}.}

\item{sedc_bandwidth}{numeric(1).
Distance at which the source concentration is reduced to
\code{exp(-3)} (approximately -95 \%)}

\item{target_fields}{character(varying). Field names in characters.}
}
\value{
data.frame (tibble) object with input field names with
a suffix \code{"_sedc"} where the sums of EDC are stored.
Additional attributes are attached for the EDC information.
\itemize{
\item attr(result, "sedc_bandwidth"): the bandwidth where
concentration reduces to approximately five percent
\item attr(result, "sedc_threshold"): the threshold distance
at which emission source points are excluded beyond that
}
}
\description{
Calculate Sum of Exponentially Decaying Contributions (SEDC) covariates
}
\note{
The function is originally from
\href{https://github.com/Spatiotemporal-Exposures-and-Toxicology/chopin}{chopin}
Distance calculation is done with terra functions internally.
Thus, the function internally converts sf objects in
\code{point_*} arguments to terra.
The threshold should be carefully chosen by users.
}
\examples{
library(terra)
library(sf)
set.seed(101)
ncpath <- system.file("gpkg/nc.gpkg", package = "sf")
nc <- terra::vect(ncpath)
nc <- terra::project(nc, "EPSG:5070")
pnt_locs <- terra::centroids(nc, inside = TRUE)
pnt_locs <- pnt_locs[, "NAME"]
pnt_from <- terra::spatSample(nc, 100L)
pnt_from$pid <- seq(1, 100)
pnt_from <- pnt_from[, "pid"]
pnt_from$val1 <- rgamma(100L, 1, 0.05)
pnt_from$val2 <- rgamma(100L, 2, 1)

vals <- c("val1", "val2")
summarize_sedc(pnt_locs, pnt_from, "NAME", 1e5, 2e5, vals)
}
\references{
\itemize{
\item \href{https://dx.doi.org/10.1021/es203152a}{Messier KP, Akita Y, \& Serre ML. (2012). Integrating Address Geocoding, Land Use Regression, and Spatiotemporal Geostatistical Estimation for Groundwater Tetrachloroethylene. \emph{Environmental Science \& Technology} 46(5), 2772-2780.}
\item Wiesner C. (n.d.). \href{https://mserre.sph.unc.edu/BMElab_web/SEDCtutorial/index.html}{Euclidean Sum of Exponentially Decaying Contributions Tutorial}
}
}
\author{
Insang Song
}
