% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_terraclimate}
\alias{calc_terraclimate}
\title{Calculate TerraClimate covariates}
\usage{
calc_terraclimate(
  from = NULL,
  locs = NULL,
  locs_id = NULL,
  radius = 0,
  fun = "mean",
  geom = FALSE,
  ...
)
}
\arguments{
\item{from}{SpatRaster(1). Output from \code{process_terraclimate()}.}

\item{locs}{data.frame. character to file path, SpatVector, or sf object.}

\item{locs_id}{character(1). Column within \code{locations} CSV file
containing identifier for each unique coordinate location.}

\item{radius}{integer(1). Circular buffer distance around site locations.
(Default = 0).}

\item{fun}{character(1). Function used to summarize multiple raster cells
within sites location buffer (Default = \code{mean}).}

\item{geom}{logical(1). Should the geometry of \code{locs} be returned in the
\code{data.frame}? Default is \code{FALSE}. If \code{geom = TRUE} and \code{locs} contain
polygon geometries, the \verb{$geometry} column in the returned data frame may
make the \code{data.frame} difficult to read due to long geometry strings. The
coordinate reference system of the \verb{$geometry} is the coordinate
reference system of \code{from}.}

\item{...}{Placeholders.}
}
\value{
a data.frame object
}
\description{
Extract TerraClimate values at point locations. Returns a \code{data.frame}
object containing \code{locs_id} and TerraClimate variable. TerraClimate
variable column name reflects the TerraClimate variable and
circular buffer radius.
}
\note{
TerraClimate data has monthly temporal resolution, so the \verb{$time} column
will contain the year and month in YYYYMM format (ie. January, 2018 =
201801).
}
\seealso{
\code{\link[=process_terraclimate]{process_terraclimate()}}
}
\author{
Mitchell Manware
}
