% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_time}
\alias{calc_time}
\title{Covariate time auxiliary function}
\usage{
calc_time(time, format)
}
\arguments{
\item{time}{character. Time value.}

\item{format}{character(1). Desired output format based on time value.
Can be "timeless", "hour", "date", "year", or "yearmonth".}
}
\value{
NULL ("timeless"); POSIXt ("hour"); Date ("date"); integer
("year" and "yearmonth")
}
\description{
Convert character value to time value.
}
\keyword{auxiliary}
