% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_tri}
\alias{calc_tri}
\title{Calculate TRI covariates}
\usage{
calc_tri(
  locs,
  from = NULL,
  locs_id = "site_id",
  radius = c(1000L, 10000L, 50000L)
)
}
\arguments{
\item{locs}{sf/SpatVector. Locations where TRI variables are calculated.}

\item{from}{SpatVector. Point vector object of TRI.
See \link{process_tri}}

\item{locs_id}{character(1). Unique site identifier column name.
Default is \code{"site_id"}.}

\item{radius}{Circular buffer radius.
Default is \code{c(1000, 10000, 50000)} (meters)}
}
\value{
A data.frame object.
}
\description{
Calculate TRI covariates
}
\note{
U.S. context.
}
\seealso{
\link{calc_sedc}
}
\author{
Insang Song, Mariana Kassien
}
