% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates_auxiliary.R
\name{calc_weighted_fun}
\alias{calc_weighted_fun}
\title{Resolve weighted summary function names for exactextractr}
\usage{
calc_weighted_fun(fun, weighted = FALSE)
}
\arguments{
\item{fun}{character(1)}

\item{weighted}{logical(1)}
}
\value{
character(1)
}
\description{
Resolve weighted summary function names for exactextractr
}
\keyword{auxiliary}
\keyword{internal}
