% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calculate_edgar}
\alias{calculate_edgar}
\title{Calculate EDGAR covariates}
\usage{
calculate_edgar(
  from,
  locs,
  locs_id = "site_id",
  radius = 0,
  weights = NULL,
  .by_time = NULL,
  geom = FALSE,
  ...
)
}
\arguments{
\item{from}{SpatRaster(1). Output from \code{process_edgar()}.}

\item{locs}{data.frame, character to file path, SpatVector, or sf object.}

\item{locs_id}{character(1). Column within \code{locations} CSV file containing
identifier for each unique coordinate location.}

\item{radius}{numeric(1). Circular buffer distance around site locations.
Default is \code{0}.}

\item{weights}{\code{NULL}, \code{SpatRaster}, polygon \code{SpatVector}/\code{sf}, or file
path. Optional weights raster for weighted extraction. If \code{NULL}
(default), unweighted extraction is performed.}

\item{.by_time}{NULL or character(1). Optional time grouping key used
with \code{.by_time} for temporal summaries.}

\item{geom}{FALSE/"sf"/"terra".. Should the function return with geometry?
Default is \code{FALSE}, options with geometry are "sf" or "terra". The
coordinate reference system of the \code{sf} or \code{SpatVector} is that of \code{from.}}

\item{...}{Placeholders.}
}
\value{
a data.frame or SpatVector object
}
\description{
Extract EDGAR gridded emissions values at point locations. For
\code{radius = 0}, cell values are extracted directly. For \code{radius > 0},
means are calculated over a circular buffer around each location.
}
\examples{
## NOTE: Example is wrapped in `\dontrun{}` as function requires data that is
##       not included in the package.
\dontrun{
loc <- data.frame(id = "001", lon = -78.90, lat = 35.97)
calculate_edgar(
  from = edgar, # derived from process_edgar() example
  locs = loc,
  locs_id = "id",
  radius = 0,
  geom = FALSE
)
}
}
\seealso{
\code{\link[=process_edgar]{process_edgar()}}
}
\author{
Mariana Alifa Kassien, Insang Song
}
