% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_auxiliary.R
\name{check_destfile}
\alias{check_destfile}
\title{Check if destination file exists or is 0 bytes.}
\usage{
check_destfile(destfile)
}
\arguments{
\item{destfile}{character(1). Destination file path.}
}
\value{
logical(1)
}
\description{
Check if destination file exists or is 0 bytes. If either condition is
met, the function returns \code{TRUE} to allow the download to proceed.
}
\keyword{auxiliary}
\keyword{internal}
