% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_auxiliary.R
\name{check_file_size}
\alias{check_file_size}
\title{Compare file sizes}
\usage{
check_file_size(url, file)
}
\arguments{
\item{url}{character(1). URL of data file to be downloaded.}

\item{file}{character(1). Destination path of the data file to be
downloaded.}
}
\value{
logical object
}
\description{
Compare the size of a locally stored data file to the size of the to-be
downloaded file, retrieved with \code{httr2}. This check helps to ensure that
incomplete or corrupted data files are re-downloaded if the file path
currently exists.
}
\author{
Mitchell Manware
}
\keyword{auxiliary}
