% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates_auxiliary.R
\name{check_geom}
\alias{check_geom}
\title{Check that \code{geom} value is one of \code{FALSE}, \code{"sf"}, or \code{"terra"}}
\usage{
check_geom(geom)
}
\arguments{
\item{geom}{FALSE/"sf"/"terra".'}
}
\value{
NULL; will stop if \code{geom} is not one of the three options
}
\description{
Check that \code{geom} value is one of \code{FALSE}, \code{"sf"},
or \code{"terra"}.
}
\author{
Mitchell Manware
}
\keyword{auxiliary}
\keyword{internal}
