% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{check_mysf}
\alias{check_mysf}
\title{Check \code{sf} object}
\usage{
check_mysf(x)
}
\arguments{
\item{x}{an \code{sf} object}
}
\value{
NULL; returns stop error if \code{x} does not match class
and/or column expectations
}
\description{
Check an \code{sf} object for its class, \code{$geometry} column, and
geometry class.
}
\author{
Eva Marques
}
\keyword{internal}
\keyword{spacetime}
