% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{check_mysftime}
\alias{check_mysftime}
\title{Check if the sftime object is formated on a specific way}
\usage{
check_mysftime(x)
}
\arguments{
\item{x}{a sftime object}
}
\description{
Check if the sftime object is formated on a specific way
}
\author{
Eva Marques
}
