% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_auxiliary.R
\name{check_urls}
\alias{check_urls}
\title{Implement \code{check_url_status}}
\usage{
check_urls(urls = urls, size = NULL, method = c("HEAD", "GET", "SKIP"))
}
\arguments{
\item{urls}{character vector of URLs}

\item{size}{number of observations to be sampled from \code{urls}}

\item{method}{httr method to obtain URL (\code{"HEAD"} or \code{"GET"}). If set to
\code{"SKIP"}, the HTTP status will not be checked and returned.}
}
\value{
logical vector for URL status = 200
}
\description{
Apply \code{check_url_status()} function to a sample of download URLs.
}
\keyword{auxiliary}
