% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates_auxiliary.R
\name{collapse_nlcd}
\alias{collapse_nlcd}
\title{Collapse listed NLCD values while filling in NA for sites outside data.}
\usage{
collapse_nlcd(
  data,
  mode = c("terra", "extract"),
  locs = NULL,
  locs_id = "site_id"
)
}
\arguments{
\item{data}{Buffered values from NLCD data.}

\item{mode}{"exact" or "terra"}

\item{locs}{extraction locations.}

\item{locs_id}{character(1). Name of unique identifier.}
}
\description{
Collapse listed NLCD values while filling in NA for sites outside data.
}
\keyword{auxiliary}
\keyword{internal}
