% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{convert_stdt}
\alias{convert_stdt}
\title{Convert a stdt to sf/sftime/SpatVector}
\usage{
convert_stdt(
  stdt,
  class_to = c("sf", "sftime", "SpatVector", "SpatRasterDataset")
)
}
\arguments{
\item{stdt}{A stdt object}

\item{class_to}{character(1). Should be one of
\code{"sf"}, \code{"sftime"}, \code{"SpatRasterDataset"}, or \code{"SpatVector"}}
}
\value{
a sf/sftime/SpatRasterDataset/SpatVector
}
\description{
Convert a stdt to sf/sftime/SpatVector
}
\author{
Insang Song
}
