% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{convert_stdt_sftime}
\alias{convert_stdt_sftime}
\title{Convert a stdtobj to sftime}
\usage{
convert_stdt_sftime(stdt)
}
\arguments{
\item{stdt}{A stdt object}
}
\value{
a sftime object
}
\description{
Convert a stdtobj to sftime
}
\author{
Eva Marques
}
