% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{convert_stdt_spatvect}
\alias{convert_stdt_spatvect}
\title{Convert a stdt to SpatVector}
\usage{
convert_stdt_spatvect(stdt)
}
\arguments{
\item{stdt}{A stdt object}
}
\value{
a SpatVector
}
\description{
Convert a stdt to SpatVector
}
\author{
Eva Marques
}
