% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{convert_stobj_to_stdt}
\alias{convert_stobj_to_stdt}
\title{Convert spatio-temporal object to a datatable with lon, lat, time, predictors columns. It also returns the crs.}
\usage{
convert_stobj_to_stdt(stobj)
}
\arguments{
\item{stobj}{object containing space-time data. It can be a data.frame,
a data.table, an sf or sftime, a SpatVector or a SpatRastDataset.}
}
\value{
a list with a "stdt" a data.table of locations identified by
lat, lon, time columns and "crs_dt" the crs of the data in well-known text
format.
}
\description{
Convert spatio-temporal object to a datatable with lon, lat, time, predictors columns. It also returns the crs.
}
\author{
Eva Marques, Insang Song
}
