% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cov}
\alias{cov}
\title{Calculate code coverage of the \code{beethoven} package with the
\code{container.sif} container.}
\usage{
cov()
}
\value{
NULL; Prints the output of the code coverage.
}
\description{
Calculate code coverage of the \code{beethoven} package with the
\code{container.sif} container.
}
\seealso{
\code{\link[covr:package_coverage]{covr::package_coverage()}}; \code{\link[covr:coverage_to_list]{covr::coverage_to_list()}}
}
\keyword{Miscellaneous}
