% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_ecoregion_data}
\alias{download_ecoregion_data}
\title{Download Ecoregion Shapefiles from EPA}
\usage{
download_ecoregion_data(
  directory_to_download = "./input/ecoregions/",
  directory_to_save = "./input/ecoregions/",
  data_download_acknowledgement = FALSE,
  unzip = TRUE,
  remove_zip = FALSE,
  download = FALSE,
  remove_command = TRUE,
  epa_certificate_path = system.file("extdata/cacert_gaftp_epa.pem", package = "amadeus")
)
}
\arguments{
\item{directory_to_download}{character(1). Directory to download zip file
of Ecoregion level 3 shapefiles}

\item{directory_to_save}{character(1). Directory to decompress zip files.}

\item{data_download_acknowledgement}{logical(1). By setting \code{TRUE} the
user acknowledge that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{unzip}{logical(1). Unzip zip files. Default \code{TRUE}.}

\item{remove_zip}{logical(1). Remove zip file from directory_to_download.
Default \code{FALSE}.}

\item{download}{logical(1). \code{FALSE} will generate a *.txt file
containing all download commands. By setting \code{TRUE} the function
will download all of the requested data files.}

\item{remove_command}{logical(1).
Remove (\code{TRUE}) or keep (\code{FALSE})
the text file containing download commands.}

\item{epa_certificate_path}{character(1). Path to the certificate file
for EPA DataCommons. Default is
'extdata/cacert_gaftp_epa.pem' under the package installation path.}
}
\value{
NULL;
}
\description{
The \code{download_ecoregion_data()} function accesses and downloads
Ecoregions level 3 data, where all pieces of information in the higher
levels are included.
}
\note{
In Linux systems as of December 2023, downloading Ecoregion data from
EPA Data Commons will result in certificate errors. This is bypassed by
manually identifying .crt file link in your browser by connecting to
https://gaftp.epa.gov then clicking a lock icon in the address bar.
(TO DEVELOPERS: see the comments in source code)
Currently we bundle the pre-downloaded crt and its PEM (which is accepted
in wget command) file in ./inst/extdata.
}
\author{
Insang Song
}
