% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_support.R
\name{download_epa_certificate}
\alias{download_epa_certificate}
\title{Check EPA certificate}
\usage{
download_epa_certificate(
  epa_certificate_path = "cacert_gaftp_epa.pem",
  certificate_url =
    "http://cacerts.digicert.com/DigiCertGlobalG2TLSRSASHA2562020CA1-1.crt"
)
}
\arguments{
\item{epa_certificate_path}{character(1).
Full path of a converted certificate of EPA.
Should end with \code{.pem}}

\item{certificate_url}{character(1).
URL of the original certificate.}
}
\value{
A file designated in \code{epa_certificate_path}
}
\description{
Check EPA certificate
}
\author{
Insang Song
}
\keyword{internal}
