% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_geos_cf_data}
\alias{download_geos_cf_data}
\title{Download atmospheric composition data from the NASA Global Earth Observing System (GEOS) model.}
\usage{
download_geos_cf_data(
  date_start = "2023-09-01",
  date_end = "2023-09-01",
  collection = c("aqc_tavg_1hr_g1440x721_v1", "chm_tavg_1hr_g1440x721_v1",
    "met_tavg_1hr_g1440x721_x1", "xgc_tavg_1hr_g1440x721_x1",
    "chm_inst_1hr_g1440x721_p23", "met_inst_1hr_g1440x721_p23"),
  directory_to_save = "./input/data/geos_cf/",
  data_download_acknowledgement = FALSE,
  download = FALSE,
  remove_command = FALSE
)
}
\arguments{
\item{date_start}{character(1). length of 10. Start date for downloading
data. Format YYYY-MM-DD (ex. September 1, 2023 = \code{"2023-09-01"}).}

\item{date_end}{character(1). length of 10. End date for downloading data.
Format YYYY-MM-DD (ex. September 1, 2023 = \code{"2023-09-01"}).}

\item{collection}{character(1). GEOS-CF data collection file name.}

\item{directory_to_save}{character(1). Directory to save data.}

\item{data_download_acknowledgement}{logical(1). By setting \code{TRUE} the
user acknowledge that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{download}{logical(1). \code{FALSE} will generate a *.txt file
containing all download commands. By setting \code{TRUE} the function
will download all of the requested data files.}

\item{remove_command}{logical(1).
Remove (\code{TRUE}) or keep (\code{FALSE})
the text file containing download commands.}
}
\value{
NULL;
}
\description{
The \code{download_goes_cf_data()} function accesses and downloads various
atmospheric composition collections from the \href{https://gmao.gsfc.nasa.gov/GEOS_systems/}{NASA Global Earth Observing System (GEOS) model}.
}
\author{
Mitchell Manware, Insang Song
}
