% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_gmted_data}
\alias{download_gmted_data}
\title{Download global elevation data from the Global Multi-resolution Terrain Elevation Data (GMTED2010).}
\usage{
download_gmted_data(
  statistic = c("Breakline Emphasis", "Systematic Subsample", "Median Statistic",
    "Minimum Statistic", "Mean Statistic", "Maximum Statistic",
    "Standard Deviation Statistic"),
  resolution = c("7.5 arc-seconds", "15 arc-seconds", "30 arc-seconds"),
  directory_to_download = "./input/gmted/",
  directory_to_save = "./input/gmted/",
  data_download_acknowledgement = FALSE,
  unzip = TRUE,
  remove_zip = FALSE,
  download = FALSE,
  remove_command = FALSE
)
}
\arguments{
\item{statistic}{character(1). Available statistics include \code{"Breakline Emphasis"}, \code{"Systematic Subsample"}, \code{"Median Statistic"},
\code{"Minimum Statistic"}, \code{"Mean Statistic"}, \code{"Maximum Statistic"}, and
\code{"Standard Deviation Statistic"}.}

\item{resolution}{character(1). Available resolutions include \code{"7.5 arc-seconds"}, \code{"15 arc-seconds"}, and \code{"30 arc-seconds"}.}

\item{directory_to_download}{character(1). Directory to download zip files
from Global Multi-resolution Terrain Elevation Data (GMTED2010).}

\item{directory_to_save}{character(1). Directory to decompress zip files.}

\item{data_download_acknowledgement}{logical(1). By setting \code{TRUE} the
user acknowledge that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{unzip}{logical(1). Unzip zip files. Default is \code{TRUE}.}

\item{remove_zip}{logical(1). Remove zip file from directory_to_download.
Default is \code{FALSE}.}

\item{download}{logical(1). \code{FALSE} will generate a *.txt file
containing all download commands. By setting \code{TRUE} the function
will download all of the requested data files.}

\item{remove_command}{logical(1).
Remove (\code{TRUE}) or keep (\code{FALSE})
the text file containing download commands. Default is FALSE.}
}
\value{
NULL;
}
\description{
The \code{download_gmted_data()} function acesses and downloads Global
Multi-resolution Terrain Elevation Data (GMTED2010) from
\href{https://www.usgs.gov/coastal-changes-and-impacts/gmted2010}{U.S. Geological Survey and National Geospatial-Intelligence Agency}.
}
\author{
Mitchell Manware, Insang Song
}
