% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_gridmet_data}
\alias{download_gridmet_data}
\title{Download gridMET data}
\usage{
download_gridmet_data(
  variables = NULL,
  year_start = 2022,
  year_end = 2022,
  directory_to_save = NULL,
  acknowledgement = FALSE,
  download = FALSE,
  remove_command = FALSE
)
}
\arguments{
\item{variables}{character(1). Variable(s) name(s). See \href{https://www.climatologylab.org/wget-gridmet.html}{gridMET Generate Wget File}
for variable names and acronym codes. (Note: variable "Burning Index" has code "bi" and variable
"Energy Release Component" has code "erc").}

\item{year_start}{integer(1). length of 4. Start of year range for
downloading data.}

\item{year_end}{integer(1). length of 4. End of year range for downloading
data.}

\item{directory_to_save}{character(1). Directory(s) to save downloaded data
files.}

\item{acknowledgement}{logical(1). By setting \code{TRUE} the
user acknowledges that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{download}{logical(1). \code{FALSE} will generate a *.txt file
containing all download commands. By setting \code{TRUE} the function
will download all of the requested data files.}

\item{remove_command}{logical(1).
Remove (\code{TRUE}) or keep (\code{FALSE})
the text file containing download commands.}
}
\value{
NULL; Yearly netCDF (.nc) files will be stored in a variable-specific
folder within \code{directory_to_save}.
}
\description{
The \code{download_gridmet_data} function accesses and downloads gridded surface meteorological data from the \href{https://www.climatologylab.org/gridmet.html}{University of California Merced Climatology Lab's gridMET dataset}.
}
\author{
Mitchell Manware
}
