% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_imperviousness}
\alias{download_imperviousness}
\title{Download urban imperviousness (developed urban descriptor for Alaska) data}
\usage{
download_imperviousness(
  collection = "Coterminous United States",
  year = 2021,
  directory_to_save = NULL,
  acknowledgement = FALSE,
  download = FALSE,
  remove_command = FALSE,
  unzip = TRUE,
  remove_zip = FALSE
)
}
\arguments{
\item{collection}{character(1). \code{"Coterminous United States"} or \code{"Alaska"}.}

\item{year}{integer(1). Available years for Coterminous United States
include \code{2001}, \code{2004}, \code{2006}, \code{2008}, \code{2011}, \code{2013}, \code{2016},
\code{2019}, and \code{2021}.
Available years for Alaska include \code{2001}, \code{2011}, and \code{2016}.}

\item{directory_to_save}{character(1). Directory to save data. Two
sub-directories will be created for the downloaded zip files ("/zip_files")
and the unzipped shapefiles ("/data_files").}

\item{acknowledgement}{logical(1). By setting \code{TRUE} the
user acknowledges that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{download}{logical(1). \code{FALSE} will generate a *.txt file
containing all download commands. By setting \code{TRUE} the function
will download all of the requested data files.}

\item{remove_command}{logical(1).
Remove (\code{TRUE}) or keep (\code{FALSE})
the text file containing download commands.}

\item{unzip}{logical(1). Unzip zip files. Default is \code{TRUE}.}

\item{remove_zip}{logical(1). Remove zip files from directory_to_download.
Default is \code{FALSE}.}
}
\value{
NULL; Zip and/or data files will be downloaded and stored in
respective sub-directories within \code{directory_to_save}.
}
\description{
The \code{download_imperviousness()} function accesses and downloads
imperviousness data from the
\href{https://www.mrlc.gov/data}{Multi-Resolution Land Characteristics (MRLC) Consortium's National Urban Imperviousness Database products data base}.
}
\references{
Dewitz, J., and U.S. Geological Survey, 2021, National Land Cover Database (NLCD) 2019 Products (ver. 2.0, June 2021): U.S. Geological Survey data release, https://doi.org/10.5066/P9KZCM54

Xian, G., Homer, C., Dewitz, J., Fry, J., Hossain, N., and Wickham, J., 2011. The change of impervious surface area between 2001 and 2006 in the conterminous United States. Photogrammetric Engineering and Remote Sensing, Vol. 77(8): 758-762.

Yang, L., Huang, C., Homer, C., Wylie, B., and M. Coan. 2003. An approach for mapping large-area impervious surfaces: Synergistic use of Landsat 7 ETM+ and high spatial resolution imagery. Canadian Journal of Remote Sensing, Vol. 29, No. 2, pp.230-240.
}
\author{
Eva Marques, Mitchell Manware, Insang Song
}
