% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_lcz}
\alias{download_lcz}
\title{Download Local Climate Zone data}
\usage{
download_lcz(
  directory_to_save = NULL,
  acknowledgement = FALSE,
  download = FALSE,
  remove_command = FALSE
)
}
\arguments{
\item{directory_to_save}{character(1). Directory to save data. Two
sub-directories will be created for the downloaded zip files ("/zip_files")
and the unzipped shapefiles ("/data_files").}

\item{acknowledgement}{logical(1). By setting \code{TRUE} the
user acknowledges that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{download}{logical(1). \code{FALSE} will generate a *.txt file
containing all download commands. By setting \code{TRUE} the function
will download all of the requested data files.}

\item{remove_command}{logical(1).
Remove (\code{TRUE}) or keep (\code{FALSE})
the text file containing download commands.}

\item{unzip}{logical(1). Unzip zip files. Default is \code{TRUE}.}

\item{remove_zip}{logical(1). Remove zip files from directory_to_download.
Default is \code{FALSE}.}
}
\value{
NULL; Zip and/or data files will be downloaded and stored in
respective sub-directories within \code{directory_to_save}.
}
\description{
The \code{download_lcz()} function accesses and downloads
Local Climate Zone data produced within produced within
\href{https://figshare.com/articles/dataset/CONUS-wide_LCZ_map_and_Training_Areas/11416950}{CONUS-wide LCZ map and Training areas}. When using the data, please cite data authors listed in References section below.
}
\references{
Demuzere et al. (2020). \emph{Combining expert and crowd-sourced training data to map urban form and functions for the continental US.} Nature Scientific Data. https://doi.org/10.1038/s41597-020-00605-z
}
\author{
Eva Marques, Mitchell Manware, Insang Song
}
