% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_modis_data}
\alias{download_modis_data}
\title{Download MODIS product files}
\usage{
download_modis_data(
  date_start = "2023-09-01",
  date_end = "2023-09-01",
  product = c("MOD09GA", "MOD11A1", "MOD06_L2", "MCD19A2", "MOD13A2", "VNP46A2"),
  version = "61",
  horizontal_tiles = c(7, 13),
  vertical_tiles = c(3, 6),
  nasa_earth_data_token = NULL,
  directory_to_save = "./input/modis/raw/",
  data_download_acknowledgement = FALSE,
  mod06_links = NULL,
  download = TRUE,
  remove_command = FALSE
)
}
\arguments{
\item{date_start}{character(1). length of 10. Start date for downloading
data. Format YYYY-MM-DD (ex. September 1, 2023 is \code{"2023-09-01"}).}

\item{date_end}{character(1). length of 10. End date for downloading data.
Format YYYY-MM-DD (ex. September 1, 2023 is \code{"2023-09-01"}).}

\item{product}{character(1).
One of \code{c("MOD09GA", "MOD11A1", "MOD06_L2", "MCD19A2", "MOD13A2", "VNP46A2")}}

\item{version}{character(1). Default is \code{"61"}, meaning v061.}

\item{horizontal_tiles}{integer(2). Horizontal tile numbers
\code{c({start}, {end})}. Default is \code{c(7, 13)}.}

\item{vertical_tiles}{integer(2). Vertical tile numbers
\code{c({start}, {end})}. Default is \code{c(3, 6)}.}

\item{nasa_earth_data_token}{character(1).
Token for downloading data from NASA. Should be set before
trying running the function.}

\item{directory_to_save}{character(1). Directory to save data.}

\item{data_download_acknowledgement}{logical(1). By setting \code{TRUE} the
user acknowledge that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{mod06_links}{character(1). CSV file path to MOD06_L2 download links
from NASA LPDAAC. Default is \code{NULL}.}

\item{download}{logical(1). Download data or only save wget commands.}

\item{remove_command}{logical(1). Remove (\code{TRUE}) or keep (\code{FALSE})
the text file containing download commands.}
}
\value{
NULL;
}
\description{
Need maintenance for the directory path change
in NASA EOSDIS. This function first retrieves the all hdf download links
on a certain day, then only selects the relevant tiles from the retrieved
links. Download is only done at the queried horizontal-vertical tile number
combinations. An exception is MOD06_L2 product, which is produced
every five minutes every day.
}
\note{
\code{date_start} and \code{date_end} should be in the same year.
Directory structure looks like
input/modis/raw/\{version\}/\{product\}/\{year\}/\{day_of_year\}
Please note that \code{date_start} and \code{date_end} are
ignored if \code{product == 'MOD06_L2'}.
}
\author{
Mitchell Manware, Insang Song
}
