% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_narr_monolevel}
\alias{download_narr_monolevel}
\title{Download meteorological data (monolevel)}
\usage{
download_narr_monolevel(
  variables = NULL,
  year_start = 2022,
  year_end = 2022,
  directory_to_save = NULL,
  acknowledgement = FALSE,
  download = FALSE,
  remove_command = FALSE
)
}
\arguments{
\item{variables}{character. Variable(s) name acronym. See \href{https://ftp.cpc.ncep.noaa.gov/NARR/fixed/merged_land_AWIP32corrected.pdf}{List of Variables in NARR Files}
for variable names and acronym codes.}

\item{year_start}{integer(1). length of 4. Start of year range for
downloading data.}

\item{year_end}{integer(1). length of 4. End of year range for downloading
data.}

\item{directory_to_save}{character(1). Directory(s) to save downloaded data
files.}

\item{acknowledgement}{logical(1). By setting \code{TRUE} the
user acknowledges that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{download}{logical(1). \code{FALSE} will generate a *.txt file
containing all download commands. By setting \code{TRUE} the function
will download all of the requested data files.}

\item{remove_command}{logical(1).
Remove (\code{TRUE}) or keep (\code{FALSE})
the text file containing download commands.}
}
\value{
NULL; netCDF (.nc) files will be stored in a variable-specific
folder within \code{directory_to_save}.
}
\description{
The \code{download_narr_monolevel} function accesses and downloads monolevel meteorological data from \href{https://psl.noaa.gov/data/gridded/data.narr.html}{NOAA's North American Regional Reanalysis (NARR) model}. "Monolevel" variables contain a single value for the entire atmospheric column (ie. Variable: Convective cloud cover; Level: Entire atmosphere considered as a single layer), or represent a specific altitude associated with the variable (ie. Variable: Air temperature; Level: 2 m).
}
\author{
Mitchell Manware, Insang Song
}
