% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_narr_p_levels}
\alias{download_narr_p_levels}
\title{Download meteorological data (pressure levels)}
\usage{
download_narr_p_levels(
  variables = NULL,
  year_start = 2022,
  year_end = 2022,
  directory_to_save = NULL,
  acknowledgement = FALSE,
  download = FALSE,
  remove_command = FALSE
)
}
\arguments{
\item{variables}{character. Variable(s) name acronym. See \href{https://ftp.cpc.ncep.noaa.gov/NARR/fixed/merged_land_AWIP32corrected.pdf}{List of Variables in NARR Files}
for variable names and acronym codes.}

\item{year_start}{integer(1). length of 4. Start of year range for
downloading data.}

\item{year_end}{integer(1). length of 4. End of year range for downloading
data.}

\item{directory_to_save}{character(1). Directory(s) to save downloaded data
files.}

\item{acknowledgement}{logical(1). By setting \code{TRUE} the
user acknowledges that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{download}{logical(1). \code{FALSE} will generate a *.txt file
containing all download commands. By setting \code{TRUE} the function
will download all of the requested data files.}

\item{remove_command}{logical(1).
Remove (\code{TRUE}) or keep (\code{FALSE})
the text file containing download commands.}
}
\value{
NULL; Monthly netCDF (.nc) files will be stored in
\code{directory_to_save}.
}
\description{
The \code{download_narr_p_levels} function accesses and downloads pressure levels meteorological data from \href{https://psl.noaa.gov/data/gridded/data.narr.html}{NOAA's North American Regional Reanalysis (NARR) model}. "Pressure levels" variables contain variable values at 29 atmospheric levels, ranging from 1000 hPa to 100 hPa. All pressure levels data will be downloaded for each variable.
}
\author{
Mitchell Manware, Insang Song
}
