% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_nei_data}
\alias{download_nei_data}
\title{Download data from EPA National Emission Inventory aggregated on-road emission data}
\usage{
download_nei_data(
  directory_to_save = "./input/nei/",
  data_download_acknowledgement = FALSE,
  download = FALSE,
  remove_command = FALSE,
  year_target = c(2017L, 2020L),
  unzip = FALSE,
  epa_certificate_path = system.file("extdata/cacert_gaftp_epa.pem", package = "amadeus")
)
}
\arguments{
\item{directory_to_save}{character(1). Directory to download files.}

\item{data_download_acknowledgement}{logical(1). By setting \code{TRUE} the
user acknowledge that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{download}{logical(1). \code{FALSE} will generate a *.txt file
containing all download commands. By setting \code{TRUE} the function
will download all of the requested data files.}

\item{remove_command}{logical(1).
Remove (\code{TRUE}) or keep (\code{FALSE})
the text file containing download commands.}

\item{year_target}{Available years of NEI data.
Default is \code{c(2017L, 2020L)}.}

\item{unzip}{logical(1). Unzip the downloaded zip files.
Default is \code{FALSE}.}

\item{epa_certificate_path}{character(1). Path to the certificate file
for EPA DataCommons. Default is
'extdata/cacert_gaftp_epa.pem' under the package installation path.}
}
\value{
NULL; Two comma-separated value (CSV) raw files for 2017 and 2020
}
\description{
Download data from EPA National Emission Inventory aggregated on-road emission data
}
\author{
Ranadeep Daw, Insang Song
}
