% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_noaa_hms_smoke_data}
\alias{download_noaa_hms_smoke_data}
\title{Download daily wildfire smoke plume data from NOAA Hazard Mapping System Fire and Smoke Product}
\usage{
download_noaa_hms_smoke_data(
  date_start = "2023-09-01",
  date_end = "2023-09-01",
  data_format = "Shapefile",
  directory_to_download = "./input/noaa_hms/",
  directory_to_save = "./input/noaa_hms/",
  data_download_acknowledgement = FALSE,
  download = FALSE,
  remove_command = FALSE,
  unzip = TRUE,
  remove_zip = FALSE
)
}
\arguments{
\item{date_start}{character(1). length of 10. Start date for downloading
data. Format YYYY-MM-DD (ex. September 1, 2023 is \code{"2023-09-01"}).}

\item{date_end}{character(1). length of 10. End date for downloading data.
Format YYYY-MM-DD (ex. September 10, 2023 is \code{"2023-09-10"}).}

\item{data_format}{character(1). "Shapefile" or "KML".}

\item{directory_to_download}{character(1). Directory to download zip files
from NOAA Hazard Mapping System Fire and Smoke Product. (Ignored if
\code{data_format = "KML"}.)}

\item{directory_to_save}{character(1). Directory to save unzipped shapefiles
and KML files.}

\item{data_download_acknowledgement}{logical(1).
By setting \code{TRUE} the
user acknowledge that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{download}{logical(1). \code{FALSE} will generate a *.txt file
containing all download commands. By setting \code{TRUE} the function
will download all of the requested data files.}

\item{remove_command}{logical(1).
Remove (\code{TRUE}) or keep (\code{FALSE})
the text file containing download commands.}

\item{unzip}{logical(1). Unzip zip files. Default is \code{TRUE}. (Ignored
if \code{data_format = "KML"}.)}

\item{remove_zip}{logical(1). Remove zip files from
directory_to_download. Default is \code{FALSE}.
(Ignored if \code{data_format = "KML"}.)}
}
\value{
NULL;
}
\description{
The \code{download_noaa_hms_smoke_data()} function accesses and downloads
wildfire smoke plume coverage data from
the National Oceanic and Atmospheric Administration's (NOAA)
\href{https://www.ospo.noaa.gov/Products/land/hms.html#0}{Hazard Mapping System Fire and Smoke Product}.
}
\author{
Mitchell Manware, Insang Song
}
