% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_olm_data}
\alias{download_olm_data}
\title{Download OpenLandMap data}
\usage{
download_olm_data(
  product = NULL,
  file_format = "tif",
  directory_to_save = NULL,
  acknowledgement = FALSE,
  download = FALSE,
  remove_command = FALSE
)
}
\arguments{
\item{product}{character(1). Available collection name in OpenLandMap
STAC Catalog. \link{list_stac_files} with \code{id_only = TRUE} to see available collections.}

\item{directory_to_save}{character(1). Directory to download files.}

\item{acknowledgement}{logical(1). By setting \code{TRUE} the
user acknowledges that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{download}{logical(1). \code{FALSE} will generate a *.txt file
containing all download commands. By setting \code{TRUE} the function
will download all of the requested data files.}

\item{remove_command}{logical(1).
Remove (\code{TRUE}) or keep (\code{FALSE})
the text file containing download commands.}
}
\value{
NULL; Yearly comma-separated value (CSV) files will be stored in
\code{directory_to_save}.
}
\description{
Accesses and downloads OpenLandMap data from the \href{https://www.openlandmap.org/}{OpenLandMap website}.
}
\note{
JSON files should be found at STAC catalog of OpenLandMap
}
\seealso{
\link{list_stac_files}
}
\author{
Insang Song
}
