% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_auxiliary.R
\name{download_remove_command}
\alias{download_remove_command}
\title{Remove download commands}
\usage{
download_remove_command(commands_txt = NULL, remove = FALSE)
}
\arguments{
\item{commands_txt}{character(1). Path of download commands}

\item{remove}{logical(1). Remove (\code{TRUE}) or
keep (\code{FALSE}) commands}
}
\description{
Remove or retain the .txt file storing all download commands.
}
\keyword{internal}
