% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_auxiliary.R
\name{download_remove_zips}
\alias{download_remove_zips}
\title{Remove zip files}
\usage{
download_remove_zips(remove = FALSE, download_name)
}
\arguments{
\item{remove}{logical(1). Confirm removal. Default is FALSE.}

\item{download_name}{character. Full zip file path}
}
\value{
NULL; removes downloaded zip files after they are unzipped
}
\description{
Remove downloaded ".zip" files.
}
\note{
!!! USE THE FUNCTION WITH CAUTION !!!
If \code{remove = TRUE}, ensure that \code{unzip = TRUE}. Choosing to remove
".zip" files without unzipping will retain none of the downloaded data.
then it will remove all files in the second higher level directory.
}
\keyword{internal}
