% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_support.R
\name{download_sanitize_path}
\alias{download_sanitize_path}
\title{Sanitize path to end with a forward slash}
\usage{
download_sanitize_path(directory)
}
\arguments{
\item{directory}{character(1). Path}
}
\value{
character ending with a forward slash.
}
\description{
Sanitize path to end with a forward slash
}
