% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_auxiliary.R
\name{download_setup_dir}
\alias{download_setup_dir}
\title{Setup directory}
\usage{
download_setup_dir(directory, zip = FALSE)
}
\arguments{
\item{directory}{character(1) directory path}

\item{zip}{logical(1). Should sub-directories be created for zip files and
data files? If \code{TRUE}, a vector of sub-directoy names will be returned.}
}
\value{
NULL; if \code{zip = TRUE} a vector of directories for zip files and
data files
}
\description{
Create \code{directory} if it does not already exist.

If directory does not exist, the directory
will be created.
}
\keyword{internal}
