% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_auxiliary.R
\name{download_sink}
\alias{download_sink}
\title{Sink download commands}
\usage{
download_sink(command_txt)
}
\arguments{
\item{command_txt}{character(1). file path to export commands.}
}
\value{
NULL; creates and opens connection to text file to store
download commands
}
\description{
Open connection to \code{command_txt} file to store download commands.
}
\keyword{internal}
