% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_tree_canopy_cover}
\alias{download_tree_canopy_cover}
\title{Download tree canopy cover data}
\usage{
download_tree_canopy_cover(
  collection = "Coterminous United States",
  year = 2021,
  directory_to_save = NULL,
  acknowledgement = FALSE,
  download = FALSE,
  remove_command = FALSE,
  unzip = TRUE,
  remove_zip = FALSE
)
}
\arguments{
\item{collection}{character(1). \code{"Coterminous United States"} or \code{"Alaska"}.}

\item{year}{integer(1). Available years for Coterminous United States
include \code{2011} to \code{2021} for both collections.}

\item{directory_to_save}{character(1). Directory to save data. Two
sub-directories will be created for the downloaded zip files ("/zip_files")
and the unzipped shapefiles ("/data_files").}

\item{acknowledgement}{logical(1). By setting \code{TRUE} the
user acknowledges that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{download}{logical(1). \code{FALSE} will generate a *.txt file
containing all download commands. By setting \code{TRUE} the function
will download all of the requested data files.}

\item{remove_command}{logical(1).
Remove (\code{TRUE}) or keep (\code{FALSE})
the text file containing download commands.}

\item{unzip}{logical(1). Unzip zip files. Default is \code{TRUE}.}

\item{remove_zip}{logical(1). Remove zip files from directory_to_download.
Default is \code{FALSE}.}
}
\value{
NULL; Zip and/or data files will be downloaded and stored in
respective sub-directories within \code{directory_to_save}.
}
\description{
The \code{download_tree_canopy_cover()} function accesses and downloads
tree canopy cover data from the
\href{https://www.mrlc.gov/data}{Multi-Resolution Land Characteristics (MRLC) Consortium's National Database products data base}.
}
\author{
Eva Marques, Mitchell Manware, Insang Song
}
