% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_tri_data}
\alias{download_tri_data}
\title{Download data from EPA toxic release inventory}
\usage{
download_tri_data(
  year_start = 2018L,
  year_end = 2022L,
  directory_to_save = "./input/tri/",
  data_download_acknowledgement = FALSE,
  download = FALSE,
  remove_command = FALSE
)
}
\arguments{
\item{year_start}{integer(1). length of 4. Start year for downloading data.}

\item{year_end}{integer(1). length of 4. End year for downloading data.}

\item{directory_to_save}{character(1). Directory to download files.}

\item{data_download_acknowledgement}{logical(1). By setting \code{TRUE} the
user acknowledge that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{download}{logical(1). \code{FALSE} will generate a *.txt file
containing all download commands. By setting \code{TRUE} the function
will download all of the requested data files.}

\item{remove_command}{logical(1). Remove (\code{TRUE}) or keep (\code{FALSE})
the text file containing download commands.}
}
\value{
NULL; Yearly comma-separated value (CSV) raw files for each year
}
\description{
Download data from EPA toxic release inventory
}
\author{
Mariana Kassien, Insang Song
}
