% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_auxiliary.R
\name{download_unzip}
\alias{download_unzip}
\title{Extract downloaded archives}
\usage{
download_unzip(file_name, directory_to_unzip, unzip = TRUE)
}
\arguments{
\item{file_name}{character(1). Full archive file path}

\item{directory_to_unzip}{character(1). Directory to extract
data}

\item{unzip}{logical(1). Extract (\code{TRUE}) or not.}
}
\value{
NULL; extracts downloaded archive files
}
\description{
Extract downloaded ".zip" or ".7z" files.
}
\keyword{internal}
