% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{dt_to_mysftime}
\alias{dt_to_mysftime}
\title{Create a sftime from a datatable}
\usage{
dt_to_mysftime(x, lonname, latname, timename, crs)
}
\arguments{
\item{x}{a data.table}

\item{lonname}{character for longitude column name}

\item{latname}{character for latitude column name}

\item{timename}{character for time column name}

\item{crs}{coordinate reference system}
}
\description{
Create a sftime from a datatable
}
\author{
Eva Marques
}
