% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{dt_to_sf}
\alias{dt_to_sf}
\title{Create a sf object from a data.table}
\usage{
dt_to_sf(datatable, crs)
}
\arguments{
\item{datatable}{A data.table object with columns "lat", "lon"}

\item{crs}{A character containing the original crs}
}
\value{
an sf object
}
\description{
Create a sf object from a data.table
}
\author{
Eva Marques
}
