% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_auxiliary.R
\name{extract_urls}
\alias{extract_urls}
\title{Extract download URLs}
\usage{
extract_urls(commands = commands, position = NULL)
}
\arguments{
\item{commands}{character vector containing download commands}

\item{position}{URL position in the vector}
}
\value{
character vector containing download URLs
}
\description{
Extract download URLs from multi-argument download commands.
}
\keyword{internal}
