% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_auxiliary.R
\name{generate_date_sequence}
\alias{generate_date_sequence}
\title{Generate date sequence}
\usage{
generate_date_sequence(date_start, date_end, sub_hyphen = TRUE)
}
\arguments{
\item{date_start}{character(1). Beginning of date sequence.}

\item{date_end}{character(1). End of date sequence.}

\item{sub_hyphen}{logical(1). Substitute hyphen in dates. If \code{TRUE}, returns
date sequence as "YYYYMMDD". If \code{FALSE}, returns date sequence as
"YYYY-MM-DD".}
}
\value{
vector
}
\description{
Generate a sequence of dates from \code{date_start} to \code{date_end}.
}
\keyword{auxiliary}
\keyword{internal}
