% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_support.R
\name{generate_time_sequence}
\alias{generate_time_sequence}
\title{Generate time sequence based on GEOS-CF data collection.}
\usage{
generate_time_sequence(collection)
}
\arguments{
\item{collection}{character(1). GEOS-CF data collection}
}
\value{
vector
}
\description{
Generate time sequence based on GEOS-CF data collection.
}
