% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{interactive}
\alias{interactive}
\title{Open interactive session with \code{container.sif} container.}
\usage{
interactive(dir = ".")
}
\arguments{
\item{dir}{character(1). Directory with \code{interactive.sh}}
}
\description{
Open interactive session with \code{container.sif} container.
}
\keyword{Miscellaneous}
