% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw.R
\name{is_date_proper}
\alias{is_date_proper}
\title{Check input strings conform to the required format}
\usage{
is_date_proper(instr = NULL, format = "\%Y-\%m-\%d")
}
\arguments{
\item{instr}{character(1). String to check.}

\item{format}{character(1). Matching format to be checked.
Default is \code{"\%Y-\%m-\%d"}, which can detect \verb{"\%Y/\%m/\%d}.
See \link{strftime} for details of formatting this string.}
}
\value{
No returning value. It stops the function if \code{instr} doesn't
conform to the \code{format}.
}
\description{
Check input strings conform to the required format
}
\author{
Insang Song
}
