% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{is_stdt}
\alias{is_stdt}
\title{Boolean to know if an object correspond to a stdtobj}
\usage{
is_stdt(obj)
}
\arguments{
\item{obj}{an object}
}
\value{
a boolean to know if obj is from newly created class "stdt"
}
\description{
Boolean to know if an object correspond to a stdtobj
}
\author{
Eva Marques
}
