% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_auxiliary.R
\name{list_stac_files}
\alias{list_stac_files}
\title{Retrieve file links from SpatioTemporal Assets Catalog (STAC)}
\usage{
list_stac_files(
  stac_json = "https://s3.eu-central-1.wasabisys.com/stac/openlandmap/catalog.json",
  format = "tif",
  which = NULL,
  id_only = FALSE
)
}
\arguments{
\item{stac_json}{character(1). Full path of STAC JSON file.}

\item{format}{character(1). Format of target files. Default is "tif".}

\item{which}{numeric/character. Index or name of collection to retrieve.}

\item{id_only}{logical(1). Return collection IDs only.}
}
\value{
character vector of file links.
}
\description{
Read file links from SpatioTemporal Assets Catalog (STAC) JSON file.
}
\note{
Retrieving URLs may take a while depending on the spatial tiling,
temporal resolution, and other assets. Users are encouraged to use
\code{which} parameter to select a specific collection.
}
\examples{
\dontrun{
read_stac_json()
}
}
\author{
Insang Song
}
\keyword{auxiliary}
