% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates_support.R
\name{modis_aggregate_sds}
\alias{modis_aggregate_sds}
\title{Aggregate layers in a MODIS sub-dataset}
\usage{
modis_aggregate_sds(
  path,
  product = c("MOD11A1", "MOD13A2", "MOD09GA", "MCD19A2"),
  nsds,
  fun_agg = "mean"
)
}
\arguments{
\item{path}{character(1). Full path to MODIS HDF4/HDF5 file.
Direct sub-dataset access is supported, for example,
HDF4_EOS:EOS_GRID:\{filename\}:\{base_grid_information\}:\{sub-dataset\}}

\item{product}{character(1). Name of MODIS product.}

\item{nsds}{character(1). Exact or regular expression filter of sub-dataset.
See \link{modis_prefilter_sds} for details.}

\item{fun_agg}{character(1). Function name to aggregate layers.
Should be acceptable to \link[terra:tapp]{terra::tapp}.}
}
\description{
Aggregate layers in a MODIS sub-dataset
}
\note{
HDF values are read as original without scaling.
Users should consult MODIS product documentation to apply proper
scaling factor for post-hoc adjustment. If users have no preliminary
information about MODIS HDF sub-datasets, consider running
\verb{terra::describe(__filename__, sds = TRUE)} to navigate the full
list of sub-datasets in the input file.
}
\seealso{
\link[terra:tapp]{terra::tapp}, \link[terra:rast]{terra::rast}, \link[terra:describe]{terra::describe}
}
\author{
Insang Song
}
