% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates_support.R
\name{modis_get_vrt}
\alias{modis_get_vrt}
\title{Get mosaicked or merged raster from multiple MODIS hdf files}
\usage{
modis_get_vrt(
  paths,
  product = c("MOD11A1", "MOD13A2", "MOD09GA", "MCD19A2"),
  date_in = NULL,
  regex_sds = NULL,
  foo = "mean"
)
}
\arguments{
\item{paths}{character. Full list of hdf file paths.
preferably a recursive search result from \code{list.files}.}

\item{product}{character(1). Product code of MODIS. Should be one of
\code{c("MOD11A1", "MOD13A2", "MOD09GA", "MCD19A2")}}

\item{date_in}{character(1). date to query. Should be in
\code{"YYYY-MM-DD"} format.}

\item{regex_sds}{character(1). subdataset names to extract.
Should conform to regular expression. See \link{regex} for details.
Default is NULL, which means that the subdataset names are automatically
selected based on \code{product} value.}

\item{foo}{Function name or custom function to aggregate overlapping
cell values. See \code{fun} description in \link[terra]{tapp} for details.}
}
\value{
A SpatRaster object.
}
\description{
Get mosaicked or merged raster from multiple MODIS hdf files
}
\author{
Insang Song
}
