% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates_support.R
\name{modis_mosaic_mod06}
\alias{modis_mosaic_mod06}
\title{Mosaic MODIS MOD06_L2 product files}
\usage{
modis_mosaic_mod06(
  paths,
  date_in,
  get_var = c("Cloud_Fraction_Day", "Cloud_Fraction_Night"),
  resolution = 0.025
)
}
\arguments{
\item{paths}{character. Full paths of hdf files.}

\item{date_in}{character(1). Date to query.}

\item{get_var}{character. One of \code{"Cloud_Fraction_Day"} or
\code{"Cloud_Fraction_Night"}}

\item{resolution}{numeric(1). Resolution of output raster.
Unit is degree.}
}
\value{
SpatRaster object. CRS is \code{"EPSG:4326"}.
}
\description{
This function will return a SpatRaster object with
mosaicked 5-minute cloud coverage values. Swath data include curvilinear
grids, which require warping/rectifying the original curvilinear grids
into rectilinear grids. The function internally warps each of inputs
then mosaic the warped images into one large SpatRaster object.
}
\author{
Insang Song
}
