% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates_support.R
\name{modis_prefilter_sds}
\alias{modis_prefilter_sds}
\title{Selected MODIS product subdataset name selector}
\usage{
modis_prefilter_sds(
  product = c("MOD11A1", "MOD13A2", "MOD09GA", "MCD19A2"),
  custom_sel = NULL
)
}
\arguments{
\item{product}{character(1). Product code.}

\item{custom_sel}{character(1). Custom filter.
If this value is not NULL, preset filter is
overridden.}
}
\value{
A character object that conforms to the regular
expression. Details of regular expression in R can be found in \link{regexp}.
}
\description{
Selected MODIS product subdataset name selector
}
\seealso{
\link{calc_modis}
}
\author{
Insang Song
}
