% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates_support.R
\name{modis_preprocess_vnp46}
\alias{modis_preprocess_vnp46}
\title{Assign MODIS VNP46 corner coordinates to retrieve a merged raster}
\usage{
modis_preprocess_vnp46(paths, date_in, subdataset = 3L, crs_ref = "EPSG:4326")
}
\arguments{
\item{paths}{character. Full paths of h5 files.}

\item{date_in}{character(1). Date to query.}

\item{subdataset}{integer(1). Subdataset number to process.
Default is 3L.}

\item{crs_ref}{character(1). terra::crs compatible CRS.
Default is "EPSG:4326"}
}
\description{
This function will return a SpatRaster object with
georeferenced h5 files of VNP46A2 product. Referencing corner coordinates
are necessary as the original h5 data do not include such information.
}
\author{
Insang Song
}
