% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates_support.R
\name{modis_warp_stars}
\alias{modis_warp_stars}
\title{Warp MODIS Swath data into rectilinear grid raster}
\usage{
modis_warp_stars(path, cellsize = 0.25, threshold = 0.5, crs_out = 4326)
}
\arguments{
\item{path}{File path of MODIS swath with exact sub-dataset specification.}

\item{cellsize}{numeric(1). Cell size (spatial resolution) of
output rectilinear grid raster.}

\item{threshold}{numeric(1). Maximum distance to fill gaps if occur.}

\item{crs_out}{integer(1)/character(1). Coordinate system definition.
Should be compatible with EPSG codes or WKT2.
See \link[terra:crs]{terra::crs} and \link[sf:st_crs]{sf::st_crs} / \href{https://www.epsg.io}{EPSG}}
}
\value{
stars object.
}
\description{
Swath data is a type of MODIS data organization,
where curvilinear points are stored with varying resolution depending on
the relative position of the sensor axis. As this type of data
typically does not work well with planar spatial data, users
should warp or rectify this data into a rectilinear raster.
Main procedure is done with \link[stars:st_warp]{stars::st_warp}, in which users are able to
customize the threshold to fill potential gaps that appear where
the target resolution is finer than the local resolution of curvilinear
grid points.
}
\note{
Users should specify sub-dataset with all flags that are
compatible with \code{gdalinfo}
}
\seealso{
\link[terra:rectify]{terra::rectify}
}
\author{
Insang Song
}
