% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_auxiliary.R
\name{narr_variable}
\alias{narr_variable}
\title{Sort NOAA NARR variables}
\usage{
narr_variable(variable)
}
\arguments{
\item{variable}{character(1). User-selected NARR variable}
}
\value{
list with URL base and vector of months (blank for monolevel)
}
\description{
Determine whether a NOAA NARR variable selected for download is a
monolevel or pressure level variable. Monolevel variables are derived
from https://downloads.psl.noaa.gov/Datasets/NARR/Dailies/monolevel/,
and pressure level variables are derived from
https://downloads.psl.noaa.gov//Datasets/NARR/Dailies/pressure/.
}
\keyword{auxiliary}
\keyword{internal}
