% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw.R
\name{process_conformity}
\alias{process_conformity}
\title{Check input assumptions}
\usage{
process_conformity(locs, check_time = FALSE)
}
\arguments{
\item{locs}{\link[=convert_stobj_to_stdt]{stdt}, \link[sf:st_as_sf]{sf},
\link[terra:vect]{SpatVector}, or \link{data.frame}}
}
\value{
SpatVector object.
}
\description{
Check if all of \code{"lon"}, \code{"lat"}, and \code{"time"}
(only if \code{check_time = TRUE}) then convert inputs into a
SpatVector object.
}
\author{
Insang Song
}
