% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_cropscape}
\alias{process_cropscape}
\title{Process CropScape data}
\usage{
process_cropscape(path = NULL, year = 2021, ...)
}
\arguments{
\item{path}{character giving CropScape data path}

\item{year}{numeric giving the year of CropScape data used}

\item{...}{Placeholders.}
}
\value{
a \code{SpatRaster} object
}
\description{
This function imports and cleans raw CropScape data,
returning a single \code{SpatRaster} object.

Reads CropScape file of selected \code{year}.
}
\author{
Insang Song
}
