% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_ecoregion}
\alias{process_ecoregion}
\title{Process ecoregion data}
\usage{
process_ecoregion(path = NULL)
}
\arguments{
\item{path}{character(1). Path to Ecoregion Shapefiles}
}
\value{
a SpatVector object
}
\description{
The \code{process_ecoregion()} function imports and cleans raw ecoregion
data, returning a SpatVector object.
}
\author{
Insang Song
}
