% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_edgar}
\alias{process_edgar}
\title{Process EDGAR emissions data}
\usage{
process_edgar(path = NULL, extent = NULL, ...)
}
\arguments{
\item{path}{character. Directory containing extracted EDGAR raster files or
one or more file paths.}

\item{extent}{numeric(4) or SpatExtent giving the extent of the raster;
if \code{NULL} (default), the entire raster is loaded.}

\item{...}{Placeholders.}
}
\value{
a \code{SpatRaster} object
}
\description{
The \code{process_edgar()} function imports extracted EDGAR gridded emissions
files and returns a single \code{SpatRaster} object. Raster formats supported by
\code{terra::rast()} such as NetCDF (\code{.nc}, \code{.nc4}) and GeoTIFF (\code{.tif},
\code{.tiff}) are supported.
}
\note{
\code{process_edgar()} currently supports gridded raster outputs from
\code{download_edgar()} such as the default \code{format = "nc"}. Plain-text EDGAR
downloads should be re-downloaded as raster outputs before processing.
}
\examples{
## NOTE: Example is wrapped in `\dontrun{}` as function requires data that is
##       not included in the package.
\dontrun{
edgar <- process_edgar(
  path = "./data/edgar",
  extent = c(-85, -75, 33, 37)
)
}
}
\seealso{
\code{\link[=download_edgar]{download_edgar()}}, \code{\link[=calculate_edgar]{calculate_edgar()}}
}
\author{
Mariana Alifa Kassien, Insang Song
}
