% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_gmted_codes}
\alias{process_gmted_codes}
\title{Generate statistic and resolution codes based on GMTED statistic and
resolution.}
\usage{
process_gmted_codes(
  string,
  statistic = FALSE,
  resolution = FALSE,
  invert = FALSE
)
}
\arguments{
\item{string}{character(1). File path to GEOS-CF data file.}

\item{statistic}{logical(1). Matches statistic to statistic code.}

\item{resolution}{logical(1). Matches resolution to resolution code.}

\item{invert}{logical(1). Default = FALSE. \code{invert = TRUE} assumes \code{string}
provides statistic or resolution code, and returns full length statistic
or resolution.}
}
\value{
character
}
\description{
Generate statistic and resolution codes based on GMTED statistic and
resolution.
}
