% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw.R
\name{process_koppen_geiger}
\alias{process_koppen_geiger}
\title{Process Koeppen-Geiger climate data}
\usage{
process_koppen_geiger(
  path = "./input/koppen_geiger/raw/Beck_KG_V1_present_0p0083.tif"
)
}
\arguments{
\item{path}{character(1). Path to Koppen-Geiger
climate zone raster file}
}
\value{
SpatRaster object
}
\description{
Process Koeppen-Geiger climate data
}
\author{
Insang Song
}
