% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_locs_radius}
\alias{process_locs_radius}
\title{Create circular buffer around site points.}
\usage{
process_locs_radius(locs, radius)
}
\arguments{
\item{locs}{SpatVector(1). SpatVector object with point geometry}

\item{radius}{integer(1). Circular buffer size (meters).}
}
\value{
SpatVector.
}
\description{
Creates a circular buffer around points if \code{radius} is > 0.
Returns points if \code{radius} is 0.
}
