% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_auxiliary.R
\name{process_merra2_time}
\alias{process_merra2_time}
\title{Process MERRA2 time steps}
\usage{
process_merra2_time(collection, from)
}
\arguments{
\item{collection}{character(1). MERRA2 collection name.}

\item{from}{SpatRaster(1). Object to extract time values from.}
}
\value{
character
}
\description{
Identify the time step of data observations based on MERRA2 collection and
filter to time values in \code{from}.
}
\keyword{auxiliary}
