% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_merra2_timestep}
\alias{process_merra2_timestep}
\title{Process MERRA2 time steps}
\usage{
process_merra2_timestep(collection)
}
\arguments{
\item{collection}{character(1). MERRA2 collection name.}
}
\value{
character
}
\description{
Identify the time step of data observations based on MERRA2 collection.
}
\keyword{internal}
