% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_modis_swath}
\alias{process_modis_swath}
\title{Mosaic MODIS swaths}
\usage{
process_modis_swath(
  path = NULL,
  date = NULL,
  subdataset = NULL,
  suffix = ":mod06:",
  resolution = 0.025,
  ...
)
}
\arguments{
\item{path}{character. Full paths of hdf files.}

\item{date}{character(1). Date to query.}

\item{subdataset}{character. Subdatasets to process.}

\item{suffix}{character(1). Should be formatted \verb{:\{product\}:},
e.g., \verb{:mod06:}}

\item{resolution}{numeric(1). Resolution of output raster.
Unit is degree.}

\item{...}{For internal use.}
}
\value{
a \code{SpatRaster} object (crs = \code{"EPSG:4326"})
}
\description{
This function will return a \code{SpatRaster} object with
values of selected subdatasets. Swath data include curvilinear
grids, which require warping/rectifying the original curvilinear grids
into rectilinear grids. The function internally warps each of inputs
then mosaic the warped images into one large \code{SpatRaster} object.
Users need to select a subdataset to process. The full path looks like
\code{"HDF4_EOS:EOS_SWATH:{file_path}:mod06:subdataset"}, where file_path is
the full path to the hdf file.
}
\seealso{
\itemize{
\item \code{\link{process_modis_warp}}
\item \href{https://gdal.org/drivers/raster/hdf4.html}{GDAL HDF4 driver documentation}
\item \code{\link[terra:describe]{terra::describe}}: to list the full subdataset list with \code{sds = TRUE}
}
}
\author{
Insang Song
}
