% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_narr}
\alias{process_narr}
\title{Import and clean NOAA NCEP North American Regional Reanalysis (NARR) data
downloaded with \code{download_narr} or \code{download_data(dataset_name = "NARR")}.
Function returns a SpatRast object containing the user-defined variable
of interest. Layer names indicate the variable, pressure level, and date
(YYYYMMDD).}
\usage{
process_narr(
  date_start = "2023-09-01",
  date_end = "2023-09-01",
  variable = NULL,
  path = "../../data/covariates/narr/"
)
}
\arguments{
\item{date_start}{character(1). length of 10. Format "YYYY-MM-DD".}

\item{date_end}{character(1). length of 10. Format "YYYY-MM-DD".}

\item{variable}{character(1). NARR variable name(s).}

\item{path}{character(1). Directory with downloaded GEOS-CF
netCDF files.}
}
\value{
a SpatRaster object
}
\description{
Import and clean NOAA NCEP North American Regional Reanalysis (NARR) data
downloaded with \code{download_narr} or \code{download_data(dataset_name = "NARR")}.
Function returns a SpatRast object containing the user-defined variable
of interest. Layer names indicate the variable, pressure level, and date
(YYYYMMDD).
}
\author{
Mitchell Manware
}
