% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R, R/process_raw.R
\name{process_nei}
\alias{process_nei}
\title{Prepare National Emission Inventory CSV files}
\usage{
process_nei(path = "./input/nei/", year = 2017, auxfile)

process_nei(path = "./input/nei/", year = 2017, auxfile)
}
\arguments{
\item{path}{character(1). Path to the directory with NEI CSV files}

\item{year}{integer(1). Data year.
Currently only accepts \code{c(2017, 2020)}}

\item{auxfile}{character(1) or SpatVector.
Path to or SpatVector object of the county boundaries in \code{year}}

\item{county}{SpatVector/sf. County boundaries.}
}
\value{
SpatVector object.

SpatVector object. The third column stores the total emission.
}
\description{
NEI data comprises multiple csv files where emissions of
50+ pollutants are recorded at county level. With raw data files,
this function will join a combined table of NEI data and county
boundary, then perform a spatial join to target locations.

NEI data comprises multiple csv files where emissions of
50+ pollutants are recorded at county level in ten EPA regions.
This function will return a combined table of NEI data.
}
\note{
Base files for \code{county} argument can be downloaded directly from
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{U.S. Census Bureau}
or by using \code{tigris} package.
}
\author{
Insang Song

Insang Song, Ranadeep Daw
}
