% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_olm}
\alias{process_olm}
\title{Process OpenLandMap data}
\usage{
process_olm(path = NULL, extent = NULL, ...)
}
\arguments{
\item{path}{character giving OpenLandMap data path}

\item{extent}{numeric(4) or SpatExtent giving the extent of the raster
if \code{NULL} (default), the entire raster is loaded}

\item{...}{Placeholders.}
}
\value{
a \code{SpatRaster} object
}
\description{
Process OpenLandMap data
}
\examples{
\dontrun{
olm <- process_olm(
  path = paste0(
    "./data/no2_s5p.l3.trop.tmwm.p50_p90_2km_a_",
    "20180501_20221130_go_epsg.4326_v20221219.tif"
 )
)
}
}
\author{
Insang Song
}
