% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_sedac_codes}
\alias{process_sedac_codes}
\title{Generate resolution codes based on NASA SEDAC population resolution.}
\usage{
process_sedac_codes(string, invert = FALSE)
}
\arguments{
\item{string}{character(1). Resolution name or code.}

\item{invert}{logical(1). Default = FALSE. \code{invert = TRUE} assumes \code{string}
provides resolution code, and returns full length resolution.}
}
\description{
Generate resolution codes based on NASA SEDAC population resolution.
}
