% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_sedac_groads}
\alias{process_sedac_groads}
\title{Process roads data}
\usage{
process_sedac_groads(path = NULL, ...)
}
\arguments{
\item{path}{character(1). Path to geodatabase or shapefiles.}

\item{...}{Placeholders.}
}
\value{
a \code{SpatVector} boject
}
\description{
The \code{process_sedac_groads()} function imports and cleans raw road data,
returning a single \code{SpatVector} object.
}
\note{
U.S. context.
}
\author{
Insang Song
}
