% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_sedac_population}
\alias{process_sedac_population}
\title{Import and clean population data from NASA Socioeconomic Data and
Applications Center (SEDAC).}
\usage{
process_sedac_population(path = "./input/sedac_population/raw/")
}
\arguments{
\item{path}{character(1). Path to GeoTIFF or netCDF file.}
}
\value{
a SpatRaster object;
}
\description{
Import and clean population data from NASA Socioeconomic Data and
Applications Center (SEDAC).
}
\author{
Mitchell Manware
}
