% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_terraclimate_codes}
\alias{process_terraclimate_codes}
\title{Process terraClimate variable codes}
\usage{
process_terraclimate_codes(string, invert = FALSE)
}
\arguments{
\item{string}{character(1). terraClimate variable name or variable code.}

\item{invert}{logical(1). Default = FALSE. \code{invert = TRUE} assumes \code{string}
provides variable code and returns full length variable name.}
}
\value{
character
}
\description{
Convert terraClimate variable names to/from variable codes.
}
\keyword{auxiliary}
