% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_auxiliary.R
\name{process_variable_codes}
\alias{process_variable_codes}
\title{Filter gridMET and terraClimate variable names and variable codes}
\usage{
process_variable_codes(variables, source = c("gridmet", "terraclimate"))
}
\arguments{
\item{variables}{character(1). Data variables. (Passed from download_* or
process_*).}

\item{source}{character(1). Data source for selected variables ("gridMET" or
"TerraClimate").}
}
\value{
character
}
\description{
Check user defined variables for gridMET and TerraClimate functions.
}
\keyword{auxiliary}
