% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{project_dt}
\alias{project_dt}
\title{Project coordinates in a datatable from crs_ori to crs_dest}
\usage{
project_dt(datatable, crs_ori, crs_dest)
}
\arguments{
\item{datatable}{A data.table object with columns "lat", "lon"}

\item{crs_ori}{A character containing the original crs of spatial data}

\item{crs_dest}{A character containing the destination crs of spatial data}
}
\value{
same datatable object with "lon", "lat",
"lon_ori", "lat_ori" columns
}
\description{
Project coordinates in a datatable from crs_ori to crs_dest
}
\note{
This function assumes that users have point geometry.
}
\author{
Eva Marques
}
