% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_test_support.R
\name{read_commands}
\alias{read_commands}
\title{Import download commands}
\usage{
read_commands(commands_path = commands_path)
}
\arguments{
\item{commands_path}{file path with wget/curl commands}
}
\value{
character vector containing download commands
}
\description{
Read download commands from .txt file and convert to character vector.
}
\keyword{internal}
