% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{rename_time}
\alias{rename_time}
\title{Rename \code{$time}}
\usage{
rename_time(x, newname)
}
\arguments{
\item{x}{an \code{sftime} object}

\item{newname}{character for new time column name}
}
\value{
an \code{sftime} object
}
\description{
Rename the \code{$time} column of an \code{sftime} object.
}
\author{
Eva Marques
}
\keyword{spacetime}
