% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{sf_as_mysftime}
\alias{sf_as_mysftime}
\title{Convert an \code{sf} to an \code{sftime}}
\usage{
sf_as_mysftime(x, timename)
}
\arguments{
\item{x}{an \code{sf} object}

\item{timename}{character: name of time column in x}
}
\value{
an \code{sftime} object
}
\description{
Convert an \code{sf} object to an \code{sftime} object. \code{x} must
contain a time-defining column, identified in \code{timename}.
}
\author{
Eva Marques
}
