% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{sftime_as_sf}
\alias{sftime_as_sf}
\title{Convert an \code{sftime} to an \code{sf}}
\usage{
sftime_as_sf(x, keeptime = TRUE)
}
\arguments{
\item{x}{an \code{sftime} object}

\item{keeptime}{boolean: TRUE if user wants to keep time column
as simple column (default = TRUE)}
}
\value{
an \code{sf} object
}
\description{
Convert an \code{sftime} object to an \code{sf} object. \code{x} must
contain a time-defining column, identified in \code{timename}.
}
\author{
Eva Marques
}
