% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{sftime_as_spatraster}
\alias{sftime_as_spatraster}
\title{Convert sftime object to SpatRaster
/!\ can be very time consuming if sftime is not spatially structured}
\usage{
sftime_as_spatraster(x, varname)
}
\arguments{
\item{x}{a sftime}

\item{varname}{variable to rasterize}
}
\value{
a SpatRaster with layers corresponding to timestamps
}
\description{
Convert sftime object to SpatRaster
/!\ can be very time consuming if sftime is not spatially structured
}
\author{
Eva Marques
}
\keyword{spacetime}
