% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{sftime_as_spatrds}
\alias{sftime_as_spatrds}
\title{Convert sftime object to SpatVectorDataset
/!\ running time can be very long if x is not
spatially and temporally structured}
\usage{
sftime_as_spatrds(x)
}
\arguments{
\item{x}{a sftime}
}
\description{
Convert sftime object to SpatVectorDataset
/!\ running time can be very long if x is not
spatially and temporally structured
}
\author{
Eva Marques
}
