% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{sftime_as_spatvector}
\alias{sftime_as_spatvector}
\title{Convert sftime object to SpatVector}
\usage{
sftime_as_spatvector(x)
}
\arguments{
\item{x}{a sftime}
}
\value{
a terra::SpatVector
}
\description{
Convert sftime object to SpatVector
}
\author{
Eva Marques
}
\keyword{spacetime}
