% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{spatraster_as_sftime}
\alias{spatraster_as_sftime}
\title{Convert a \code{SpatRaster} to an \code{sftime}}
\usage{
spatraster_as_sftime(x, varname, timename = "time")
}
\arguments{
\item{x}{a \code{SpatRaster} object}

\item{varname}{character for variable column name in the sftime}

\item{timename}{character for time column name in the sftime
(default: "time")}
}
\value{
a \code{sftime} object
}
\description{
Convert a \code{SpatRaster} object to an \code{sftime} object. \code{x} must
contain a time-defining column, identified in \code{timename}.
}
\seealso{
\link[terra:rast]{terra::rast}
}
\author{
Eva Marques
}
\keyword{spacetime}
