% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{spatvector_as_sftime}
\alias{spatvector_as_sftime}
\title{Convert a \code{SpatVector} to an \code{sftime}}
\usage{
spatvector_as_sftime(x, timename = "time")
}
\arguments{
\item{x}{a \code{SpatVector} object}

\item{timename}{character for time column name in x
(default: "time")}
}
\value{
an \code{sftime} object
}
\description{
Convert a \code{SpatVector} object to an \code{sftime} object. \code{x} must
contain a time-defining column, identified in \code{timename}.
}
\seealso{
\link[terra:vect]{terra::vect}
}
\author{
Eva Marques
}
