% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{test}
\alias{test}
\title{Run all tests within a single file from \verb{tests/testthat/} directory
with the \code{container.sif} container.}
\usage{
test(pattern = NULL)
}
\arguments{
\item{pattern}{A regular expression to match the test file name.}
}
\value{
NULL; Prints the output of the \code{testthat} tests.
}
\description{
Run all tests within a single file from \verb{tests/testthat/} directory
with the \code{container.sif} container.
}
\seealso{
\code{\link[testthat:test_file]{testthat::test_file()}}
}
\keyword{internal}
