% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_auxiliary.R
\name{test_download_functions}
\alias{test_download_functions}
\title{Download unit tests}
\usage{
test_download_functions(
  directory_to_save = directory_to_save,
  commands_path = commands_path,
  url_status = url_status
)
}
\arguments{
\item{directory_to_save}{directory to test saving}

\item{commands_path}{file path with download commands}

\item{url_status}{logical vector for URL status = 200}
}
\value{
NULL; returns stop error if one or more tests fail
}
\description{
Implement directory, file, and download URL unit tests.
}
\keyword{internal}
