% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_auxiliary.R
\name{apply_extent}
\alias{apply_extent}
\title{Apply extent to the processed data}
\usage{
apply_extent(data, extent)
}
\arguments{
\item{data}{sf/terra object.}

\item{extent}{numeric(4). Extent to filter the data.
Should be ordered as c(xmin, xmax, ymin, ymax).}
}
\value{
sf/terra object with the extent applied.
}
\description{
User-defined extent is used to filter the data.
}
\keyword{internal}
