% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{as_mysftime}
\alias{as_mysftime}
\title{Convert to sftime object on the form adapted to beethoven code}
\usage{
as_mysftime(x, ...)
}
\arguments{
\item{x}{a data.frame, data.table, SpatVector or SpatRasterDataset}

\item{...}{if x is a data.frame or data.table: lonname, latname, timename and
crs arguments are required. If x is a sf or sftime, timename argument is
required. If x is a terra::SpatRaster, varname argument is required.}
}
\value{
a sftime object with constrained time column name
(see check_mysftime() function)
}
\description{
Convert to sftime object on the form adapted to beethoven code
}
\author{
Eva Marques
}
