% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_covariates}
\alias{calc_covariates}
\title{Calculate covariates}
\usage{
calc_covariates(
  covariate = c("modis", "koppen-geiger", "koeppen-geiger", "koppen", "koeppen", "geos",
    "dummies", "gmted", "roads", "ecoregions", "ecoregion", "hms", "noaa", "smoke",
    "gmted", "narr", "narr_monolevel", "narr_p_levels", "plevels", "monolevel",
    "p_levels", "geos", "geos_cf", "sedac_population", "population"),
  locs,
  from,
  locs_id = "site_id",
  ...
)
}
\arguments{
\item{covariate}{character(1). Covariate type.}

\item{locs}{sf/SpatVector. Unique locations. Should include
a unique identifier field named \code{locs_id}}

\item{from}{character. Single or multiple from strings.}

\item{locs_id}{character(1). Name of unique identifier.
Default is \code{"site_id"}.}

\item{...}{Arguments passed to each covariate calculation
function.}
}
\value{
Calculated covariates. Mainly data.frame object.
}
\description{
Calculate covariates
}
\seealso{
\itemize{
\item \code{calc_modis}: \code{"modis"}, \code{"MODIS"}
\item \code{calc_koppen_geiger}: \code{"koppen-geiger"}, \code{"koeppen-geiger"}, \code{"koppen"},
\item \code{calc_ecoregion}: \code{"ecoregion"}, \code{"ecoregions"}
\item \code{calc_temporal_dummies}: \code{"dummies"}
\item \code{calc_hms}: \code{"hms"}, \code{"noaa"}, \code{"smoke"}
\item \code{calc_gmted}: \code{"gmted"}
\item \code{calc_narr}: \code{"narr"}, \code{"narr_monolevel"}, \verb{"narr_p_levels}",
\code{"plevels"}, \code{"monolevel"}, \code{"p_levels"}
\item \code{calc_geos}: \code{"geos"}, \code{"geos_cf"}
\item \code{calc_sedac_population}: \code{"population"}, \code{"sedac_population"}
}
}
\author{
Insang Song
}
