% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_covariates}
\alias{calc_covariates}
\title{Calculate covariates}
\usage{
calc_covariates(
  covariate = c("modis", "koppen-geiger", "koeppen-geiger", "koppen", "koeppen", "geos",
    "dummies", "gmted", "roads", "ecoregions", "ecoregion", "hms", "noaa", "smoke",
    "gmted", "narr", "narr_monolevel", "narr_p_levels", "plevels", "monolevel",
    "p_levels", "geos", "geos_cf"),
  locs,
  from,
  locs_id = "site_id",
  ...
)
}
\arguments{
\item{covariate}{character(1). Covariate type.}

\item{locs}{sf/SpatVector. Unique locations. Should include
a unique identifier field named \code{locs_id}}

\item{from}{character. Single or multiple from strings.}

\item{locs_id}{character(1). Name of unique identifier.
Default is \code{"site_id"}.}

\item{...}{Arguments passed to each covariate calculation
function.}
}
\value{
Calculated covariates. Mainly data.frame object.
}
\description{
Calculate covariates
}
\seealso{
\itemize{
\item \code{calc_modis}: \code{"modis"}, \code{"MODIS"}
\item \code{calc_koppen_geiger}: \code{"koppen-geiger"}, \code{"koeppen-geiger"}, \code{"koppen"},
\item \code{calc_ecoregion}: \code{"ecoregion"}, \code{"ecoregions"}
\item \code{calc_temporal_dummies}: \code{"dummies"}
\item \code{calc_hms}: \code{"hms"}, \code{"noaa"}, \code{"smoke"}
\item \verb{calc_gmted``: }"gmted"`
\item \verb{calc_narr``: }"narr"\verb{, }"narr_monolevel"\verb{, }"narr_p_levels\verb{", }"plevels"\verb{, }"monolevel"\verb{, }"p_levels"`
\item \verb{calc_geos``: }"geos"\verb{, }"geos_cf"`
}
}
\author{
Insang Song
}
