% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_covariates}
\alias{calc_covariates}
\title{Calculate covariates wrapper function}
\usage{
calc_covariates(
  covariate = c("modis", "koppen-geiger", "koeppen-geiger", "koppen", "koeppen", "geos",
    "dummies", "gmted", "roads", "ecoregions", "ecoregion", "hms", "noaa", "smoke",
    "gmted", "narr", "narr_monolevel", "narr_p_levels", "plevels", "monolevel",
    "p_levels", "geos", "sedac_population", "population", "nlcd", "merra", "merra2",
    "gridmet", "terraclimate"),
  from,
  locs,
  locs_id = "site_id",
  ...
)
}
\arguments{
\item{covariate}{character(1). Covariate type.}

\item{from}{character. Single or multiple from strings.}

\item{locs}{sf/SpatVector. Unique locations. Should include
a unique identifier field named \code{locs_id}}

\item{locs_id}{character(1). Name of unique identifier.
Default is \code{"site_id"}.}

\item{...}{Arguments passed to each covariate calculation
function.}
}
\value{
Calculated covariates. Mainly data.frame object.
}
\description{
The \code{calculate_covariates()} function extracts values at point
locations from a SpatRaster or SpatVector object returned from
\code{process_covariates()}. \code{calculate_covariates()} and the underlying
source-specific covariate functions have been designed to operate on the
processed objects. To avoid errors, \strong{do not edit the processed
SpatRaster or SpatVector objects before passing to
\code{calculate_covariates()}}.
}
\note{
\code{covariate} argument value is converted to lowercase.
}
\seealso{
\itemize{
\item \code{\link{calc_modis_par}}: \code{"modis"}, \code{"MODIS"}
\item \code{\link{calc_koppen_geiger}}: \code{"koppen-geiger"}, \code{"koeppen-geiger"}, \code{"koppen"},
\item \code{\link{calc_ecoregion}}: \code{"ecoregion"}, \code{"ecoregions"}
\item 
\item \code{\link{calc_hms}}: \code{"hms"}, \code{"noaa"}, \code{"smoke"}
\item 
\item \code{\link{calc_narr}}: \code{"narr"}, \code{"narr_monolevel"}, \verb{"narr_p_levels}",
\code{"plevels"}, \code{"monolevel"}, \code{"p_levels"}
\item \code{\link{calc_geos}}: \code{"geos"}, \code{"geos_cf"}
\item \code{\link{calc_sedac_population}}: \code{"population"}, \code{"sedac_population"}
\item 
\item 
\item \code{\link{calc_merra2}}: \code{"merra"}, \code{"MERRA"}, \code{"merra2"}, \code{"MERRA2"}
\item \code{\link{calc_gridmet}}: \code{"gridMET"}, \code{"gridmet"}
\item \code{\link{calc_terraclimate}}: \code{"terraclimate"}, \code{"TerraClimate"}
}
}
\author{
Insang Song
}
