% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_ecoregion}
\alias{calc_ecoregion}
\title{Calculate EPA Ecoregions level 2/3 binary variables}
\usage{
calc_ecoregion(locs, from = NULL, locs_id = "site_id", ...)
}
\arguments{
\item{locs}{sf/SpatVector. Unique locs. Should include
a unique identifier field named \code{locs_id}}

\item{from}{SpatVector. Ecoregion polygons}

\item{locs_id}{character(1). Name of unique identifier.}

\item{...}{Placeholders.}
}
\value{
a data.frame object with dummy variables and attributes of:
\itemize{
\item \code{attr(., "ecoregion2_code")}: Ecoregion lv.2 code and key
\item \code{attr(., "ecoregion3_code")}: Ecoregion lv.3 code and key
}
}
\description{
Calculate EPA Ecoregions level 2/3 binary variables
}
\author{
Insang Song
}
