% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_ecoregion}
\alias{calc_ecoregion}
\title{Calculate EPA Ecoregions level 2/3 binary variables}
\usage{
calc_ecoregion(
  path = "./input/data/ecoregions/raw/us_eco_l3_state_boundaries.shp",
  sites,
  id_col = "site_id"
)
}
\arguments{
\item{path}{character(1). Path to Ecoregion Shapefiles}

\item{sites}{sf/SpatVector. Unique sites. Should include
a unique identifier field named \code{id_col}}

\item{id_col}{character(1). Name of unique identifier.}
}
\value{
a data.frame object with dummy variables and attributes of:
\itemize{
\item \code{attr(., "ecoregion2_code")}: Ecoregion lv.2 code and key
\item \code{attr(., "ecoregion3_code")}: Ecoregion lv.3 code and key
}
}
\description{
Calculate EPA Ecoregions level 2/3 binary variables
}
\author{
Insang Song
}
