% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_geos}
\alias{calc_geos}
\title{Calculate NASA GEOS-CF meteorological and atmospheric covariates}
\usage{
calc_geos(from, locs, locs_id = NULL, radius = 0, fun = "mean")
}
\arguments{
\item{from}{SpatRaster(1). Cleaned SpatRaster object that has been returned
from \code{process_eos} containing GEOS-CF variable data.}

\item{locs}{data.frame, characater to file path, SpatVector, or sf object.}

\item{locs_id}{character(1). Column within \code{locations} CSV file
containing identifier for each unique coordinate location.}

\item{radius}{integer(1). Circular buffer distance around site locations.
(Default = 0).}

\item{fun}{character(1). Function used to summarize multiple raster cells
within sites location buffer (Default = \code{mean}).}
}
\value{
a data.frame object;
}
\description{
Extract GEOS-CF data at point locations using SpatRaster object from
\code{process_geos}. Function returns a data frame containing GEOS-CF variable
values at user-defined sites. Unique column names reflect variable name,
circular buffer, and vertical pressure level (if applicable).
}
\author{
Mitchell Manware
}
