% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_gmted}
\alias{calc_gmted}
\title{Calculate elevation covariates}
\usage{
calc_gmted(
  from,
  locs,
  locs_id = NULL,
  radius = 0,
  fun = "mean",
  geom = FALSE,
  ...
)
}
\arguments{
\item{from}{SpatRaster(1). Output from \code{process_gmted()}.}

\item{locs}{data.frame. character to file path, SpatVector, or sf object.}

\item{locs_id}{character(1). Column within \code{locations} CSV file
containing identifier for each unique coordinate location.}

\item{radius}{integer(1). Circular buffer distance around site locations.
(Default = 0).}

\item{fun}{character(1). Function used to summarize multiple raster cells
within sites location buffer (Default = \code{mean}).}

\item{geom}{logical(1). Should the geometry of \code{locs} be returned in the
\code{data.frame}? Default is \code{FALSE}. If \code{geom = TRUE} and \code{locs} contain
polygon geometries, the \verb{$geometry} column in the returned data frame may
make the \code{data.frame} difficult to read due to long geometry strings. The
coordinate reference system of the \verb{$geometry} is the coordinate
reference system of \code{from}.}

\item{...}{Placeholders}
}
\value{
a data.frame object
}
\description{
Extract elevation values at point locations. Returns a \code{data.frame}
object containing \code{locs_id}, year of release, and elevation variable.
Elevation variable column name reflects the elevation statistic, spatial
resolution of \code{from}, and circular buffer radius (ie. Breakline Emphasis
at 7.5 arc-second resolution with 0 meter buffer: breakline_emphasis_r75_0).
}
\seealso{
\code{\link[=process_gmted]{process_gmted()}}
}
\author{
Mitchell Manware
}
