% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_gridmet}
\alias{calc_gridmet}
\title{Calculate gridMET covariates}
\usage{
calc_gridmet(from, locs, locs_id = NULL, radius = 0, fun = "mean")
}
\arguments{
\item{from}{SpatRaster(1). Output from \code{process_gridmet()}.}

\item{locs}{data.frame. character to file path, SpatVector, or sf object.}

\item{locs_id}{character(1). Column within \code{locations} CSV file
containing identifier for each unique coordinate location.}

\item{radius}{integer(1). Circular buffer distance around site locations.
(Default = 0).}

\item{fun}{character(1). Function used to summarize multiple raster cells
within sites location buffer (Default = \code{mean}).}
}
\value{
a data.frame object
}
\description{
Extract gridMET values at point locations. Returns a \code{data.frame}
object containing \code{locs_id} and gridMET variable. gridMET variable
column name reflects the gridMET variable and circular buffer radius.
}
\seealso{
\code{\link[=process_gridmet]{process_gridmet()}}
}
\author{
Mitchell Manware
}
