% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_hms}
\alias{calc_hms}
\title{Calculate wildfire smoke covariates}
\usage{
calc_hms(from, locs, locs_id = NULL, radius = 0, ...)
}
\arguments{
\item{from}{SpatVector(1). Output of \code{process_hms()}.}

\item{locs}{data.frame, characater to file path, SpatVector, or sf object.}

\item{locs_id}{character(1). Column within \code{locations} CSV file
containing identifier for each unique coordinate location.}

\item{radius}{integer(1). Circular buffer distance around site locations.
(Default = 0).}

\item{...}{Placeholders.}
}
\value{
a data.frame object
}
\description{
Extract wildfire smoke plume values at point locations. Returns a
\code{data.frame} object containing \code{locs_id}, date, and binary variable
for wildfire smoke plume density inherited from \code{from} (0 = point not
covered by wildfire smoke plume; 1 = point covered by wildfire smoke plume).
}
\seealso{
\code{\link[=process_hms]{process_hms()}}
}
\author{
Mitchell Manware
}
