% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_koppen_geiger}
\alias{calc_koppen_geiger}
\title{Calculate Koeppen-Geiger climate zone binary variables}
\usage{
calc_koppen_geiger(
  locs = NULL,
  from = NULL,
  locs_id = "site_id",
  variables = NULL
)
}
\arguments{
\item{locs}{sf/SpatVector. Unique locs. Should include
a unique identifier field named \code{locs_id}}

\item{from}{character(1). from to Koppen-Geiger
climate zone raster file}

\item{locs_id}{character(1). Name of unique identifier.}
}
\value{
a data.frame object
}
\description{
Calculate Koeppen-Geiger climate zone binary variables
}
\author{
Insang Song
}
