% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_koppen_geiger}
\alias{calc_koppen_geiger}
\title{Calculate Koeppen-Geiger climate zone binary variables}
\usage{
calc_koppen_geiger(
  path = "./input/koppen_geiger/raw/Beck_KG_V1_present_0p0083.tif",
  sites,
  id_col = "site_id"
)
}
\arguments{
\item{path}{character(1). Path to Koppen-Geiger
climate zone raster file}

\item{sites}{sf/SpatVector. Unique sites. Should include
a unique identifier field named \code{id_col}}

\item{id_col}{character(1). Name of unique identifier.}
}
\value{
a data.frame object
}
\description{
Calculate Koeppen-Geiger climate zone binary variables
}
\author{
Insang Song
}
