% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_merra2}
\alias{calc_merra2}
\title{Calculate meteorological and atmospheric covariates}
\usage{
calc_merra2(from, locs, locs_id = NULL, radius = 0, fun = "mean", ...)
}
\arguments{
\item{from}{SpatRaster(1). Output of \code{process_merra2()}.}

\item{locs}{data.frame, characater to file path, SpatVector, or sf object.}

\item{locs_id}{character(1). Column within \code{locations} CSV file
containing identifier for each unique coordinate location.}

\item{radius}{integer(1). Circular buffer distance around site locations.
(Default = 0).}

\item{fun}{character(1). Function used to summarize multiple raster cells
within sites location buffer (Default = \code{mean}).}

\item{...}{Placeholders}
}
\value{
a data.frame object
}
\description{
Extract meteorological and atmospheric values at point locations. Returns a
\code{data.frame} object containing \code{locs_id}, date and hour, vertical
pressure level, and meteorological or atmospheric variable. Variable column
name reflects variable and circular buffer radius.
}
\seealso{
\code{\link[=calc_geos]{calc_geos()}}, \code{\link[=process_merra2]{process_merra2()}}
}
\author{
Mitchell Manware
}
