% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_modis_par}
\alias{calc_modis_par}
\title{Calculate MODIS product covariates in multiple CPU threads}
\usage{
calc_modis_par(
  from = NULL,
  locs = NULL,
  locs_id = "site_id",
  radius = c(0L, 1000L, 10000L, 50000L),
  preprocess = process_modis_merge,
  name_covariates = NULL,
  subdataset = NULL,
  fun_summary = "mean",
  nthreads = floor(length(parallelly::availableWorkers())/2),
  package_list_add = NULL,
  export_list_add = NULL,
  max_cells = 1e+08,
  ...
)
}
\arguments{
\item{from}{character. List of paths to MODIS/VIIRS files.}

\item{locs}{sf/SpatVector object. Unique locs where covariates
will be calculated.}

\item{locs_id}{character(1). Site identifier. Default is \code{"site_id"}}

\item{radius}{numeric. Radii to calculate covariates.
Default is \code{c(0, 1000, 10000, 50000)}.}

\item{preprocess}{function. Function to handle HDF files.}

\item{name_covariates}{character. Name header of covariates.
e.g., \code{"MOD_NDVIF_0_"}.
The calculated covariate names will have a form of
'{name_covariates}{zero-padded buffer radius in meters}',
e.g., 'MOD_NDVIF_0_50000' where 50 km radius circular buffer
was used to calculate mean NDVI value.}

\item{subdataset}{Index or search pattern of subdataset.}

\item{fun_summary}{character or function. Function to summarize
extracted raster values.}

\item{nthreads}{integer(1). Number of threads to be used
to calculate covariates.}

\item{package_list_add}{character. A vector with package names to load
these in each thread. Note that \code{sf}, \code{terra}, \code{exactextractr},
\code{doParallel}, \code{parallelly} and \code{dplyr} are the default packages to be
loaded.}

\item{export_list_add}{character. A vector with object names to export
to each thread. It should be minimized to spare memory.}

\item{max_cells}{integer(1). Maximum number of cells to be read at once.
Higher values will expedite processing, but will increase memory usage.
Maximum possible value is \code{2^31 - 1}.
See \code{\link[exactextractr:exact_extract]{exactextractr::exact_extract}} for details.}

\item{...}{Arguments passed to \code{preprocess}.}
}
\description{
\code{calc_modis_par} essentially runs \code{\link{calc_modis_daily}} function
in each thread (subprocess). Based on daily resolution, each day's workload
will be distributed to each thread. With \code{product} argument,
the files are processed by a customized function where the unique structure
and/or characteristics of the products are considered. \code{nthreads}
argument should be carefully selected in consideration of the machine's
CPU and memory capacities as products have their own memory pressure.
\code{locs} should be \code{sf} object as it is exportable to parallel workers.
}
\note{
Overall, this function and dependent routines assume that the file
system can handle concurrent access to the (network) disk by multiple
processes. File system characteristics, package versions, and hardware
settings and specification can affect the processing efficiency.
\code{locs} is expected to be convertible to \code{sf} object. \code{sf}, \code{SpatVector}, and
other class objects that could be converted to \code{sf} can be used.
Common arguments in \code{preprocess} functions such as \code{date} and \code{path} are
automatically detected and passed to the function. Please note that
\code{locs} here and \code{path} in \code{preprocess} functions are assumed to have a
standard naming convention of raw files from NASA.
}
\seealso{
See details for setting parallelization:
\itemize{
\item \code{\link[foreach:foreach]{foreach::foreach}}
\item \code{\link[parallelly:makeClusterPSOCK]{parallelly::makeClusterPSOCK}}
\item \code{\link[parallelly:availableCores]{parallelly::availableCores}}
\item \code{\link[doParallel:registerDoParallel]{doParallel::registerDoParallel}}
}

This function leverages the calculation of single-day MODIS
covariates:
\itemize{
\item \code{\link{calc_modis_daily}}
}

Also, for preprocessing, see:
\itemize{
\item \code{\link{process_modis_merge}}
\item \code{\link{process_modis_swath}}
\item \code{\link{process_bluemarble}}
}
}
