% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_narr}
\alias{calc_narr}
\title{Calculate meteorological covariates}
\usage{
calc_narr(from, locs, locs_id = NULL, radius = 0, fun = "mean")
}
\arguments{
\item{from}{SpatRaster(1). Output of \code{process_narr()}.}

\item{locs}{data.frame, characater to file path, SpatVector, or sf object.}

\item{locs_id}{character(1). Column within \code{locations} CSV file
containing identifier for each unique coordinate location.}

\item{radius}{integer(1). Circular buffer distance around site locations.
(Default = 0).}

\item{fun}{character(1). Function used to summarize multiple raster cells
within sites location buffer (Default = \code{mean}).}
}
\value{
a data.frame object
}
\description{
Extract meteorological values at point locations. Returns a \code{data.frame}
object containing \code{locs_id}, date, vertical pressure level, and
meteorological variable. Meteorological variable column name reflects
variable and circular buffer radius.
}
\seealso{
\code{\link{process_narr}}
}
\author{
Mitchell Manware
}
