% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_nei}
\alias{calc_nei}
\title{Calculate National Emission Inventory (NEI) covariates}
\usage{
calc_nei(locs, from = NULL, locs_id = "site_id")
}
\arguments{
\item{locs}{sf/SpatVector. Locations at NEI values are joined.}

\item{from}{character(1). from to the directory with NEI CSV files}

\item{locs_id}{character(1). Unique site identifier column name.
Default is \code{"site_id"}. It is no more than a placeholder
in this function}
}
\value{
A data.frame object.
}
\description{
Calculate National Emission Inventory (NEI) covariates
}
\author{
Insang Song, Ranadeep Daw
}
