% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_nlcd}
\alias{calc_nlcd}
\title{Calculate land cover covariates}
\usage{
calc_nlcd(from, locs, locs_id = "site_id", radius = 1000, ...)
}
\arguments{
\item{from}{SpatRaster(1). Output of \code{process_nlcd()}.}

\item{locs}{terra::SpatVector of points geometry}

\item{locs_id}{character(1). Unique identifier of locations}

\item{radius}{numeric (non-negative) giving the
radius of buffer around points}

\item{...}{Placeholders.}
}
\value{
a data.frame object
}
\description{
Compute ratio of land cover class in circle buffers around points. Returns
a \code{data.frame} object containing \code{locs_id}, longitude, latitude,
time (year), and computed ratio for each land cover class.
}
\seealso{
\code{\link[=process_nlcd]{process_nlcd()}}
}
