% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_nlcd_ratio}
\alias{calc_nlcd_ratio}
\title{Compute land cover classes ratio in circle buffers around points}
\usage{
calc_nlcd_ratio(path, sites, radius = 1000, year = 2021)
}
\arguments{
\item{path}{character giving nlcd data path}

\item{sites}{terra::SpatVector of points geometry}

\item{radius}{numeric (non-negative) giving the
radius of buffer around points}

\item{year}{numeric giving the year of NLCD data used}
}
\description{
Compute land cover classes ratio in circle buffers around points
}
