% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_sedac_groads}
\alias{calc_sedac_groads}
\title{Calculate roads covariates}
\usage{
calc_sedac_groads(
  from = NULL,
  locs = NULL,
  locs_id = NULL,
  radius = 1000,
  fun = sum
)
}
\arguments{
\item{from}{SpatVector(1). Output of \code{process_sedac_groads}.}

\item{locs}{data.frame, characater to file path, SpatVector, or sf object.}

\item{locs_id}{character(1). Column within \code{locations} CSV file
containing identifier for each unique coordinate location.}

\item{radius}{integer(1). Circular buffer distance around site locations.
(Default = 1000).}

\item{fun}{function(1). Function used to summarize the length of roads
within sites location buffer (Default is \code{sum}).}
}
\value{
a data.frame object with three columns.
}
\description{
Prepared groads data is clipped with the buffer polygons
of \code{radius}. The total length of the roads are calculated.
Then the density of the roads is calculated by dividing
the total length from the area of the buffer. \code{terra::linearUnits()}
is used to convert the unit of length to meters.
}
\note{
Unit is km / sq km.
}
\seealso{
\code{\link{process_sedac_groads}}
}
\author{
Insang Song
}
