% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_temporal_dummies}
\alias{calc_temporal_dummies}
\title{Calculate temporal dummy variables}
\usage{
calc_temporal_dummies(
  locs,
  from = NULL,
  locs_id = "site_id",
  year = seq(2018L, 2022L)
)
}
\arguments{
\item{locs}{data.frame with a temporal field named \code{"time"}}

\item{from}{not used.}

\item{locs_id}{character(1). Unique site identifier column name.
Default is \code{"site_id"}.}

\item{year}{integer. Year domain to dummify.
Default is \code{seq(2018L, 2022L)}}
}
\value{
data.frame with year, month, and weekday indicators.
}
\description{
Calculate temporal dummy variables
}
\author{
Insang Song
}
