% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_temporal_dummies}
\alias{calc_temporal_dummies}
\title{Calculate temporal dummy variables}
\usage{
calc_temporal_dummies(
  sites,
  id_col = "site_id",
  domain_year = seq(2018L, 2022L)
)
}
\arguments{
\item{sites}{data.frame with a temporal field named \code{"time"}
see \code{\link{convert_stobj_to_stdt}}}

\item{id_col}{character(1). Unique site identifier column name.
Default is \code{"site_id"}.}

\item{domain_year}{integer. Year domain to dummify.
Default is \code{seq(2018L, 2022L)}}
}
\value{
data.frame with year, month, and weekday indicators.
}
\description{
Calculate temporal dummy variables
}
\author{
Insang Song
}
