% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates_auxiliary.R
\name{calc_time}
\alias{calc_time}
\title{Prepare time values}
\usage{
calc_time(time, format)
}
\arguments{
\item{time}{Time value}

\item{format}{Type of time to return in the \verb{$time} column. Can be
"timeless" (ie. Ecoregions data), "date" (ie. NARR data), "hour"
(ie. GEOS data), "year" (ie. SEDAC population data), or "yearmonth"
(ie. TerraClimate data).}
}
\value{
a \code{Date}, \code{POSIXt}, or \code{integer} object based on \verb{format =}
}
\description{
Prepare the time values for covariate calculation based on type of time
value.
}
\keyword{auxiliary}
\keyword{internal}
