% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_tri}
\alias{calc_tri}
\title{Calculate toxic release covariates}
\usage{
calc_tri(
  from = NULL,
  locs,
  locs_id = "site_id",
  radius = c(1000L, 10000L, 50000L),
  ...
)
}
\arguments{
\item{from}{SpatVector(1). Output of \code{process_tri()}.}

\item{locs}{sf/SpatVector. Locations where TRI variables are calculated.}

\item{locs_id}{character(1). Unique site identifier column name.
Default is \code{"site_id"}.}

\item{radius}{Circular buffer radius.
Default is \code{c(1000, 10000, 50000)} (meters)}

\item{...}{Placeholders.}
}
\value{
a data.frame object
}
\description{
Extract toxic release values at point locations. Returns a \code{data.frame}
object containing \code{locs_id} and variables for each chemical in
\code{from}.
}
\note{
U.S. context.
}
\seealso{
\code{\link[=calc_sedc]{calc_sedc()}}, \code{\link[=process_tri]{process_tri()}}
}
\author{
Insang Song, Mariana Kassien
}
