% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_tri}
\alias{calc_tri}
\title{Calculate TRI covariates}
\usage{
calc_tri(
  path = "./input/tri/",
  sites,
  id_col = "site_id",
  domain_year = seq(2018L, 2022L),
  radius = c(1000L, 10000L, 50000L),
  sites_epsg = "EPSG:4326"
)
}
\arguments{
\item{path}{character(1). Path to the directory with TRI CSV files}

\item{sites}{stdt/sf/SpatVector/data.frame. Unique sites
see \code{\link{convert_stobj_to_stdt}}}

\item{id_col}{character(1). Unique site identifier column name.
Default is \code{"site_id"}.}

\item{domain_year}{integer. Year domain to dummify.
Default is \code{seq(2018L, 2022L)}}

\item{radius}{Circular buffer radius.
Default is \code{c(1000, 10000, 50000)} (meters)}

\item{sites_epsg}{character(1). Coordinate system of sites.}
}
\value{
A data.frame object.
}
\description{
Calculate TRI covariates
}
\author{
Insang Song
}
