% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{check_mysf}
\alias{check_mysf}
\title{Check if the sf object is formated on a specific way}
\usage{
check_mysf(x)
}
\arguments{
\item{x}{a sf object}
}
\description{
Check if the sf object is formated on a specific way
}
\author{
Eva Marques
}
\keyword{spacetime}
