% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{check_mysftime}
\alias{check_mysftime}
\title{Check \code{sftime} object}
\usage{
check_mysftime(x)
}
\arguments{
\item{x}{an \code{sftime} object}
}
\value{
NULL; returns stop error if \code{x} does not match class
and column expectations
}
\description{
Check an \code{sftime} object for its class, \code{$time} column,
\code{$geometry} column, and geometry class.
}
\author{
Eva Marques
}
\keyword{spacetime}
