% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_auxiliary.R
\name{check_url_status}
\alias{check_url_status}
\title{Check HTTP status}
\usage{
check_url_status(url, method = c("HEAD", "GET"))
}
\arguments{
\item{url}{Download URL to be checked.}

\item{method}{httr method to obtain URL (\code{"HEAD"} or \code{"GET"})}
}
\value{
logical object
}
\description{
Check if provided URL returns HTTP status 200 or 206.
}
\author{
Insang Song; Mitchell Manware
}
\keyword{auxiliary}
