% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_test_support.R
\name{check_urls}
\alias{check_urls}
\title{Sample download URLs and apply \code{check_url_status} function}
\usage{
check_urls(urls = urls, size = NULL, method = c("HEAD", "GET"))
}
\arguments{
\item{urls}{character vector of URLs}

\item{size}{number of observations to be sampled from \code{urls}}

\item{method}{httr method to obtain URL (\code{"HEAD"} or \code{"GET"})}
}
\value{
logical vector for URL status = 200
}
\description{
Sample download URLs and apply \code{check_url_status} function
}
