% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_aqs_data}
\alias{download_aqs_data}
\title{Download daily data from AQS datamart}
\usage{
download_aqs_data(
  parameter_code = 88101,
  year_start = 2018,
  year_end = 2022,
  resolution_temporal = "daily",
  directory_to_download = "./input/aqs/",
  directory_to_save = "./input/aqs/",
  url_aqs_download = "https://aqs.epa.gov/aqsweb/airdata/",
  data_download_acknowledgement = FALSE,
  unzip = TRUE,
  remove_zip = FALSE,
  download = FALSE,
  remove_command = FALSE
)
}
\arguments{
\item{parameter_code}{integer(1). length of 5.
EPA pollutant parameter code. For details, please refer to
\href{https://aqs.epa.gov/aqsweb/documents/codetables/parameters.html}{AQS parameter codes}}

\item{year_start}{integer(1). length of 4.
Start year for downloading data.}

\item{year_end}{integer(1). length of 4.
End year for downloading data.}

\item{resolution_temporal}{character(1).
Name of column containing POC values.
Currently, no value other than \code{"daily"} works.}

\item{directory_to_download}{character(1).
Directory to download zip files from AQS data mart.}

\item{directory_to_save}{character(1).
Directory to decompress zip files.}

\item{url_aqs_download}{character(1).
URL to the AQS pre-generated datasets.}

\item{data_download_acknowledgement}{logical(1). By setting \code{TRUE} the
user acknowledge that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{unzip}{logical(1). Unzip zip files. Default \code{TRUE}.}

\item{remove_zip}{logical(1). Remove zip file from directory_to_download.
Default \code{FALSE}.}

\item{download}{logical(1). \code{FALSE} will generate a *.txt file
containing all download commands. By setting \code{TRUE} the function
will download all of the requested data files.}

\item{remove_command}{logical(1).
Remove (\code{TRUE}) or keep (\code{FALSE})
the text file containing download commands. Default is FALSE.}
}
\value{
NULL; Separate comma-separated value (CSV) files of
monitors and the daily representative values
will be stored in \code{directory_to_save}.
}
\description{
Download daily data from AQS datamart
}
\author{
Mariana Kassien, Insang Song, Mitchell Manware
}
