% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_data}
\alias{download_data}
\title{Download raw data wrapper function}
\usage{
download_data(
  dataset_name = c("aqs", "ecoregion", "geos", "gmted", "koppen", "koppengeiger",
    "merra2", "merra", "narr_monolevel", "modis", "narr_p_levels", "nlcd", "noaa",
    "sedac_groads", "sedac_population", "groads", "population", "plevels", "p_levels",
    "monolevel", "hms", "smoke", "tri", "nei", "gridmet", "terraclimate", "huc",
    "cropscape", "cdl", "prism", "olm", "openlandmap"),
  directory_to_save = NULL,
  acknowledgement = FALSE,
  ...
)
}
\arguments{
\item{dataset_name}{character(1). Dataset to download.}

\item{directory_to_save}{character(1). Directory to save / unzip
(if zip files are downloaded) data.}

\item{acknowledgement}{logical(1). By setting \code{TRUE} the
user acknowledges that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{...}{Arguments passed to each download function.}
}
\description{
The \code{download_data()} function accesses and downloads atmospheric, meteorological, and environmental data from various open-access data sources.
}
\note{
\itemize{
\item All download function names are in \code{download_*} formats
}
}
\seealso{
For details of each download function per dataset,
Please refer to:
\itemize{
\item \link{download_aqs}: "aqs", "AQS"
\item \link{download_ecoregion}: "ecoregion"
\item \link{download_geos}: "geos"
\item \link{download_gmted}: "gmted", "GMTED"
\item \link{download_koppen_geiger}: "koppen", "koppengeiger"
\item \link{download_merra2}: "merra2", "merra", "MERRA", "MERRA2"
\item \link{download_narr_monolevel}: "narr_monolevel", "monolevel"
\item \link{download_narr_p_levels}: "narr_p_levels", "p_levels", "plevels"
\item \link{download_nlcd}: "nlcd", "NLCD"
\item \link{download_hms}: "noaa", "smoke", "hms"
\item \link{download_sedac_groads}: "sedac_groads", "groads"
\item \link{download_sedac_population}: "sedac_population", "population"
\item \link{download_modis}: "modis", "MODIS"
\item \link{download_tri}: "tri", "TRI"
\item \link{download_nei}: "nei", "NEI"
\item \link{download_gridmet}: "gridMET", "gridmet"
\item \link{download_terraclimate}: "TerraClimate", "terraclimate"
}
}
\author{
Insang Song
}
