% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_data}
\alias{download_data}
\title{Download raw data wrapper function}
\usage{
download_data(
  dataset_name = c("aqs", "ecoregion", "ecoregions", "geos", "gmted", "koppen",
    "koppengeiger", "merra2", "merra", "narr_monolevel", "modis", "narr_p_levels",
    "nlcd", "noaa", "sedac_groads", "sedac_population", "groads", "population",
    "plevels", "p_levels", "monolevel", "hms", "smoke", "tri", "nei", "gridmet",
    "terraclimate", "huc", "cropscape", "cdl", "prism", "olm", "openlandmap"),
  directory_to_save = NULL,
  acknowledgement = FALSE,
  ...
)
}
\arguments{
\item{dataset_name}{character(1). Dataset to download.}

\item{directory_to_save}{character(1). Directory to save / unzip
(if zip files are downloaded) data.}

\item{acknowledgement}{logical(1). By setting \code{TRUE} the
user acknowledges that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{...}{Arguments passed to each download function.}
}
\description{
The \code{download_data()} function accesses and downloads atmospheric, meteorological, and environmental data from various open-access data sources.
}
\note{
\itemize{
\item All download function names are in \code{download_*_data} formats
}
}
\seealso{
For details of each download function per dataset,
Please refer to:
\itemize{
\item \link{`download_aqs_data`}: \code{"aqs"}, \code{"AQS"}
\item \link{`download_ecoregion_data`}: \code{"ecoregions"}, \code{"ecoregion"}
\item \link{`download_geos_data`}: \code{"geos"}
\item \link{`download_gmted_data`}: \code{"gmted"}, \code{"GMTED"}
\item \link{`download_koppen_geiger_data`}: \code{"koppen"}, \code{"koppengeiger"}
\item \link{`download_merra2_data`}: "merra2", \code{"merra"}, \code{"MERRA"}, \code{"MERRA2"}
\item \link{`download_narr_monolevel_data`}: \code{"narr_monolevel"}, \code{"monolevel"}
\item \link{`download_narr_p_levels_data`}: \code{"narr_p_levels"}, \code{"p_levels"}, \code{"plevels"}
\item \link{`download_nlcd_data`}: \code{"nlcd"}, \code{"NLCD"}
\item \link{`download_hms_data`}: \code{"noaa"}, \code{"smoke"}, \code{"hms"}
\item \link{`download_sedac_groads_data`}: \code{"sedac_groads"}, \code{"groads"}
\item \link{`download_sedac_population_data`}: \code{"sedac_population"}, \code{"population"}
\item \link{`download_modis_data`}: \code{"modis"}, \code{"MODIS"}
\item \link{`download_tri_data`}: \code{"tri"}, \code{"TRI"}
\item \link{`download_nei_data`}: \code{"nei"}, \code{"NEI"}
\item \link{`download_gridmet_data`}: \code{"gridMET"}, \code{"gridmet"}
\item \link{`download_terraclimate_data`}: \code{"TerraClimate"}, \code{"terraclimate"}
}
}
\author{
Insang Song
}
