% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_forest_canopy_height}
\alias{download_forest_canopy_height}
\title{Download forest canopy height data}
\usage{
download_forest_canopy_height(
  collection = "NAM",
  directory_to_save = NULL,
  acknowledgement = FALSE,
  download = FALSE,
  remove_command = FALSE
)
}
\arguments{
\item{collection}{character(1). \code{"NAM"}, \code{"SAM"}, \code{"NAFR"}, \code{"SAFR"},
\code{"NASIA"}, \code{"SASIA"} or \code{"AUS"}.}

\item{directory_to_save}{character(1). Directory to save data.}

\item{acknowledgement}{logical(1). By setting \code{TRUE} the
user acknowledges that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{download}{logical(1). \code{FALSE} will generate a *.txt file
containing all download commands. By setting \code{TRUE} the function
will download all of the requested data files.}

\item{remove_command}{logical(1).
Remove (\code{TRUE}) or keep (\code{FALSE})
the text file containing download commands.}
}
\value{
NULL; Zip and/or data files will be downloaded and stored in
respective sub-directories within \code{directory_to_save}.
}
\description{
The \code{download_forest_canopy_height()} function accesses and downloads
2019 forest canopy height data from the
\href{https://glad.umd.edu/dataset/gedi}{Global Land Analysis and Discovery (GLAD) laboratory Global Forest Canopy Height product}.
}
\author{
Eva Marques, Mitchell Manware, Insang Song
}
