% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_gridmet}
\alias{download_gridmet}
\title{Download gridMET data}
\usage{
download_gridmet(
  variables = NULL,
  year = c(2018, 2022),
  directory_to_save = NULL,
  acknowledgement = FALSE,
  download = FALSE,
  remove_command = FALSE
)
}
\arguments{
\item{variables}{character(1). Variable(s) name(s). See \href{https://www.climatologylab.org/wget-gridmet.html}{gridMET Generate Wget File}
for variable names and acronym codes. (Note: variable "Burning Index" has code "bi" and variable
"Energy Release Component" has code "erc").}

\item{year}{character(2). length of 4 each. Start/end years for downloading data.}

\item{directory_to_save}{character(1). Directory(s) to save downloaded data
files.}

\item{acknowledgement}{logical(1). By setting \code{TRUE} the
user acknowledges that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{download}{logical(1). \code{FALSE} will generate a *.txt file
containing all download commands. By setting \code{TRUE} the function
will download all of the requested data files.}

\item{remove_command}{logical(1).
Remove (\code{TRUE}) or keep (\code{FALSE})
the text file containing download commands.}
}
\value{
NULL; netCDF (.nc) files will be stored in a variable-specific
folder within \code{directory_to_save}.
}
\description{
The \code{download_gridmet} function accesses and downloads gridded surface meteorological data from the \href{https://www.climatologylab.org/gridmet.html}{University of California Merced Climatology Lab's gridMET dataset}.
}
\examples{
download_gridmet(
  variables = "Precipitation",
  year = c(2023, 2023),
  directory_to_save = tempdir(),
  acknowledgement = TRUE,
  download = FALSE, # NOTE: download skipped for examples,
  remove_command = TRUE
)
}
\references{
\insertRef{article_abatzoglou2013development}{amadeus}
}
\author{
Mitchell Manware
}
