% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_nei}
\alias{download_nei}
\title{Download road emissions data}
\usage{
download_nei(
  epa_certificate_path = NULL,
  certificate_url = paste0("http://cacerts.digicert.com/",
    "DigiCertGlobalG2TLSRSASHA2562020CA1-1.crt"),
  year = c(2017L, 2020L),
  directory_to_save = NULL,
  acknowledgement = FALSE,
  download = TRUE,
  remove_command = FALSE,
  unzip = TRUE,
  remove_zip = FALSE,
  show_progress = TRUE,
  hash = FALSE,
  max_tries = 20,
  rate_limit = 2
)
}
\arguments{
\item{epa_certificate_path}{TO BE DEPRECATED. Certificate path.}

\item{certificate_url}{TO BE DEPRECATED. Certificate URL.}

\item{year}{integer(1). Available years of NEI data.}

\item{directory_to_save}{character(1). Directory to save data.}

\item{acknowledgement}{logical(1). Must be TRUE to proceed.}

\item{download}{logical(1). DEPRECATED. Downloads happen automatically.}

\item{remove_command}{logical(1). Deprecated, ignored.}

\item{unzip}{logical(1). Unzip zip files (default TRUE).}

\item{remove_zip}{logical(1). Remove zip files after unzipping (default
FALSE).}

\item{show_progress}{logical(1). Show download progress (default TRUE)}

\item{hash}{logical(1). Return hash of downloaded files (default FALSE)}

\item{max_tries}{integer(1). Maximum retry attempts (default 20)}

\item{rate_limit}{numeric(1). Minimum seconds between requests (default 2)}
}
\value{
invisible list with download results; or hash character if hash=TRUE
}
\description{
The \code{download_nei()} function accesses and downloads road emissions
data from the U.S Environmental Protection Agency's (EPA) National
Emissions Inventory (NEI).
}
\note{
NEI data does not require authentication.
}
\examples{
\dontrun{
download_nei(
  year = c(2017L, 2020L),
  directory_to_save = tempdir(),
  acknowledgement = TRUE
)
}
}
\references{
\insertRef{web_usepa2024nei}{amadeus}
}
\author{
Kyle Messier, Insang Song
}
