% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_olm_data}
\alias{download_olm_data}
\title{Download OpenLandMap data}
\usage{
download_olm_data(
  product = NULL,
  format = "tif",
  directory_to_save = NULL,
  acknowledgement = FALSE,
  download = FALSE,
  remove_command = FALSE
)
}
\arguments{
\item{product}{character(1). Available collection name in OpenLandMap
STAC Catalog. \link{list_stac_files} with \code{id_only = TRUE} to see available collections.
\itemize{
\item "no2_s5p.l3.trop.tmwm"
\item "no2_s5p.l3.trop.tmwm.ltm"
\item "log.oc_iso.10694"
\item "evi_mod13q1.stl.trend.logit.ols.beta"
\item "land.cover_esacci.lc.l4"
\item "evi_mod13q1.tmwm.inpaint"
\item "dtm.bareearth_ensemble"
\item "fapar_essd.lstm"
\item "fapar_essd.lstm.p95.beta"
\item "pot.fapar_fapar.p95.eml.m"
\item "pot.fapar_fapar.p95.eml"
\item "snow.cover_esa.modis"
\item "snow.cover_esa.modis.ltm"
\item "wilderness_li2022.human.footprint"
\item "wv_mcd19a2v061.seasconv"
\item "wv_mcd19a2v061.seasconv.m_p50"
\item "wv_mcd19a2v061.seasconv.m_p25"
\item "wv_mcd19a2v061.seasconv.m_p75"
\item "wv_mcd19a2v061.seasconv.m_std"
\item "wv_mcd19a2v061.seasconv.m.yearly"
\item "bulkdens.fineearth_usda.4a1h"
\item "geom_merit.dem"
\item "fapar_proba.v"
\item "forest.cover_esacci.ifl"
\item "grtgroup_usda.soiltax"
\item "land.cover_copernicus"
\item "organic.carbon.stock_msa.kgm2"
\item "organic.carbon_usda.6a1c"
\item "pft.landcover_esa.cci.lc"
\item "precipitation_sm2rain.ltm"
\item "ph.h2o_usda.4c1a2a"
\item "pop.count_ghs.jrc"
\item "sand.wfraction_usda.3a1a1a"
\item "lc_mcd12q1v061.p1"
\item "texture.class_usda.tt"
\item "water.occurrence_jrc.surfacewater"
\item "watercontent.33kPa_usda.4b1c"
\item "dsm_glo30"
\item "lc_glad.glcluc"
\item "lc_glad.glcluc.change"
\item "landuse.cropland_hyde"
\item "landuse.pasture_hyde"
\item "land.use.land.cover_hilda.plus"
\item "lst_mod11a2.daytime.trend.logit.ols.beta"
\item "lst_mod11a2.nighttime.trend.logit.ols.beta"
\item "lst_mod11a2.daytime.annual"
\item "lst_mod11a2.nighttime.annual"
\item "lst_mod11a2.daytime"
\item "lst_mod11a2.nighttime"
\item "landform_usgs.ecotapestry"
\item "lithology_usgs.ecotapestry"
\item "grtgroup_usda.soiltax.hapludalfs"
\item "biome.type_biome00k"
\item "biomes_biome6k.tropical.evergreen.broadleaf.forest"
\item "biomes_biome6k.tropical.evergreen.broadleaf.forest.rcp26"
\item "biomes_biome6k.tropical.evergreen.broadleaf.forest.rcp45"
\item "biomes_biome6k.tropical.evergreen.broadleaf.forest.rcp85"
\item "biomes_biome6k.tropical.savanna"
\item "biomes_biome6k.tropical.savanna.rcp26"
\item "biomes_biome6k.tropical.savanna.rcp45"
\item "biomes_biome6k.tropical.savanna.rcp85"
\item "lc_glc.fcs30d"
\item "nightlights.average_viirs.v21"
\item "nightlights.difference_viirs.v21"
\item "l2a.gedi"
\item "fluxnet"
\item "gbov"
\item "geowiki.lc"
\item "geowiki.forest.loss"
\item "veg.plot"
\item "obis"
\item "fapar.eml"
}}

\item{format}{character(1). File format to query. Default is "tif".
It could be used as a pattern search for the file names.}

\item{directory_to_save}{character(1). Directory to download files.}

\item{acknowledgement}{logical(1). By setting \code{TRUE} the
user acknowledges that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{download}{logical(1). \code{FALSE} will generate a *.txt file
containing all download commands. By setting \code{TRUE} the function
will download all of the requested data files.}

\item{remove_command}{logical(1).
Remove (\code{TRUE}) or keep (\code{FALSE})
the text file containing download commands.}
}
\value{
NULL; GeoTIFF (.tif) files will be stored in
\code{directory_to_save}.
}
\description{
Accesses and downloads OpenLandMap data from the \href{https://www.openlandmap.org/}{OpenLandMap website}.
}
\note{
\code{extdata/openlandmap_assets.rds} contains the available assets in OpenLandMap.
Users may want to check the available assets to download data directly.
For developers: JSON files should be found at STAC catalog of OpenLandMap when updated.
}
\seealso{
\link{list_stac_files}
}
\author{
Insang Song
}
