% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_support.R
\name{download_permit}
\alias{download_permit}
\title{Check for data download acknowledgement}
\usage{
download_permit(data_download_acknowledgement)
}
\arguments{
\item{data_download_acknowledgement}{logical(1). Whether to
start downloading}
}
\description{
Check for data download acknowledgement
}
