% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_support.R
\name{download_remove_zips}
\alias{download_remove_zips}
\title{Remove downloaded zip files}
\usage{
download_remove_zips(remove = FALSE, download_name)
}
\arguments{
\item{remove}{logical(1). Confirm removal. Default is FALSE.}

\item{download_name}{character. Full zip file path}
}
\description{
Remove downloaded zip files
}
