% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_support.R
\name{download_setup_dir}
\alias{download_setup_dir}
\title{Check if input directory exists}
\usage{
download_setup_dir(directory)
}
\arguments{
\item{directory}{character(1) directory path}
}
\description{
If directory does not exist, the directory
will be created.
}
