% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_support.R
\name{download_sink}
\alias{download_sink}
\title{Start sink download commands into a text file}
\usage{
download_sink(command_txt)
}
\arguments{
\item{command_txt}{character(1). file path to export commands.}
}
\description{
Start sink download commands into a text file
}
