% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_support.R
\name{download_unzip}
\alias{download_unzip}
\title{Unzip downloaded data}
\usage{
download_unzip(file_name, directory_to_unzip, unzip = TRUE)
}
\arguments{
\item{file_name}{character(1). Full zip file path}

\item{directory_to_unzip}{character(1). Directory to unzip
data}

\item{unzip}{logical(1). Unzip (\code{TRUE}) or not.}
}
\description{
Unzip downloaded data
}
