% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_test_support.R
\name{extract_urls}
\alias{extract_urls}
\title{Extract URLs from download commands}
\usage{
extract_urls(commands = commands, position = NULL)
}
\arguments{
\item{commands}{character vector containing download commands}

\item{position}{URL position in the vector}
}
\value{
character vector containing download URLs
}
\description{
Extract URLs from download commands
}
