% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{interactive}
\alias{interactive}
\title{Open interactive session with \code{container.sif} container.}
\usage{
interactive(dir = "./tests/container")
}
\description{
Open interactive session with \code{container.sif} container.
}
\keyword{Miscellaneous}
