% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_aqs}
\alias{process_aqs}
\title{Process unique U.S. EPA AQS sites}
\usage{
process_aqs(
  path = NULL,
  date = c("2018-01-01", "2022-12-31"),
  mode = c("full", "sparse", "location"),
  data_field = "Arithmetic.Mean",
  return_format = c("terra", "sf", "data.table"),
  ...
)
}
\arguments{
\item{path}{character(1). Directory path to daily measurement data.}

\item{date}{character(2). Start and end date.
Should be in \code{"YYYY-MM-DD"} format and sorted. If \code{NULL},
only unique locations are returned.}

\item{mode}{character(1). One of "full" (all dates * all locations)
or "sparse" (date-location pairs with available data) or
"location" (unique locations).}

\item{data_field}{character(1). Data field to extract.}

\item{return_format}{character(1). \code{"terra"} or \code{"sf"} or \code{"data.table"}.}

\item{...}{Placeholders.}
}
\value{
a \code{SpatVector} or sf object depending on the \code{return_format}
}
\description{
The \code{process_aqs()} function cleans and imports raw air quality
monitoring sites, returning a single \code{SpatVector} or sf object.
}
\note{
Choose \code{date} and \code{mode} values with caution.
The function may return a massive data.table, resulting in
a long processing time or even a crash.
}
\seealso{
\itemize{
\item \code{\link[=download_aqs_data]{download_aqs_data()}}
\item \href{https://aqs.epa.gov/aqsweb/documents/codetables/parameters.csv}{EPA, n.d., \emph{AQS Parameter Codes}}
}
}
