% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_aqs}
\alias{process_aqs}
\title{Filter unique AQS sites with or without temporal information}
\usage{
process_aqs(
  path = NULL,
  date = c("2018-01-01", "2022-12-31"),
  return_format = "terra"
)
}
\arguments{
\item{path}{character(1). Directory path to daily measurement data.}

\item{date}{character(2). Start and end date.
Should be in \code{"YYYY-MM-DD"} format and sorted. If \code{NULL},
only unique locations are returned.}

\item{return_format}{character(1). \code{"terra"} or \code{"sf"}.}
}
\value{
SpatVector or sf object depending on the \code{return_format}.
}
\description{
Filter unique AQS sites with or without temporal information
}
\note{
\code{date = NULL} will return a massive data.table
object. Please choose proper \code{date} values.
}
