% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_aqs}
\alias{process_aqs}
\title{Filter unique AQS sites with or without temporal information}
\usage{
process_aqs(
  path = "./tests/testdata/daily_88101_2018-2022.rds",
  date = c("2018-01-01", "2022-12-31"),
  return_format = "terra"
)
}
\arguments{
\item{path}{character(1). Path to daily measurement data.
RDS and CSV are accepted.}

\item{date}{character(2). Start and end date.
Should be in \code{"YYYY-MM-DD"} format and sorted. If \code{NULL},
only unique locations are returned.}
}
\value{
data.table object with three or four fields.
\itemize{
\item "site_id"
\item "lon": in WGS 1984 (EPSG:4326)
\item "lat": in WGS 1984 (EPSG:4326)
\item "time"
"time" field will be attached only when \code{include_time = TRUE}.
}
}
\description{
Filter unique AQS sites with or without temporal information
}
\note{
\code{date = NULL} will return a massive data.table
object. Please choose proper \code{date} values.
}
\examples{
library(terra)
library(data.table)
aqs_sites <- process_aqs()
}
