% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_aqs}
\alias{process_aqs}
\title{Filter unique AQS sites with or without temporal information}
\usage{
process_aqs(
  path = "./tests/testdata/daily_88101_2018-2022.rds",
  include_time = FALSE,
  date_start = "2018-01-01",
  date_end = "2022-12-31",
  return_format = "terra"
)
}
\arguments{
\item{path}{character(1). Path to daily measurement data.
RDS and CSV are accepted.}

\item{include_time}{logical(1). Should the output include
temporal information?}

\item{date_start}{character(1). Start date.
Should be in "YYYY-MM-DD" format.}

\item{date_end}{character(1). End date.
Should be in "YYYY-MM-DD" format.}
}
\value{
data.table object with three or four fields.
\itemize{
\item "site_id"
\item "lon": in WGS 1984 (EPSG:4326)
\item "lat": in WGS 1984 (EPSG:4326)
\item "time"
"time" field will be attached only when \code{include_time = TRUE}.
}
}
\description{
Filter unique AQS sites with or without temporal information
}
\note{
\code{include_time = TRUE} will return a massive data.table
object. Please choose proper \code{date_start} and \code{date_end} values.
}
\examples{
library(terra)
library(data.table)
aqs_sites <- process_aqs()
}
