% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_collection}
\alias{process_collection}
\title{Process GEOS-CF and MERRA2 collection codes}
\usage{
process_collection(
  path,
  source,
  collection = FALSE,
  date = FALSE,
  datetime = FALSE
)
}
\arguments{
\item{path}{character(1). File path to data file.}

\item{source}{character(1). "geos" for GEOS-CF or "merra2" for MERRA2}

\item{collection}{logical(1). Identifies and returns collection
name(s) based on provided file path(s).}

\item{date}{logical(1). Identifies and returns date sequence (YYYYMMDD) based
on provided file path(s).}

\item{datetime}{logical(1). Identifies and returns date time sequence
(YYYYMoMoDDHHMiMi) based on provided file path(s).}
}
\value{
character
}
\description{
Identify the GEOS-CF or MERRA2 collection based on the file path.
}
\keyword{auxiliary}
