% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_covariates}
\alias{process_covariates}
\title{Process covariates}
\usage{
process_covariates(
  covariate = c("modis", "koppen-geiger", "koeppen-geiger", "koppen", "koeppen", "geos",
    "dummies", "gmted", "roads", "sedac_groads", "nlcd", "tri", "ncep", "nei",
    "ecoregions", "ecoregion"),
  path,
  ...
)
}
\arguments{
\item{covariate}{character(1). Covariate type.}

\item{path}{character(1). Directory or file path to raw data
depending on \code{covariate} value.}

\item{...}{Arguments passed to each raw data processing function.
function.}
}
\value{
\code{SpatVector} or \code{SpatRaster} depending on covariate type.
}
\description{
Process covariates
}
\seealso{
\itemize{
\item \code{process_modis}: \code{"modis"}, \code{"MODIS"}
\item \code{process_koppen_geiger}: \code{"koppen-geiger"}, \code{"koeppen-geiger"}, \code{"koppen"},
\item \code{process_ecoregion}: \code{"ecoregion"}, \code{"ecoregions"}
}
}
\author{
Insang Song
}
