% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw.R
\name{process_ecoregion}
\alias{process_ecoregion}
\title{Process EPA Ecoregion shapefiles}
\usage{
process_ecoregion(
  path = "./input/data/ecoregions/raw/us_eco_l3_state_boundaries.shp"
)
}
\arguments{
\item{path}{character(1). Path to Ecoregion Shapefiles}
}
\value{
SpatVector object.
}
\description{
Process EPA Ecoregion shapefiles
}
\author{
Insang Song
}
