% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_edgar}
\alias{process_edgar}
\title{Process EDGAR data}
\usage{
process_edgar(
  year = c(2018, 2021),
  path = NULL,
  extent = NULL,
  voc = 1:25,
  sector_voc = c("AGRICULTURE", "BUILDINGS", "FUEL_EXPLOITATION", "IND_COMBUSTION",
    "IND_PROCESSES", "POWER_INDUSTRY", "TRANSPORT", "WASTE"),
  ...
)
}
\arguments{
\item{year}{numeric(1 or 2). Year (1) or start and end Years (2).}

\item{path}{character(1). Directory with downloaded netCDF (.nc) files.}

\item{extent}{numeric(4) or SpatExtent giving the extent of the raster
if \code{NULL} (default), the entire raster is loaded}

\item{voc}{Integer vector. Used for VOC speciation in version
"8.1_voc". Accepts integers from 1 to 25. See:
https://edgar.jrc.ec.europa.eu/dataset_ap81_VOC_spec#p3  for reference on
speciation groups and VOC numbers.}

\item{sector_voc}{Character vector. Emission sectors for VOC speciation
data. Supported values: "AGRICULTURE", "BUILDINGS", "FUEL_EXPLOITATION",
"IND_COMBUSTION", "IND_PROCESSES", "POWER_INDUSTRY", "TRANSPORT", "WASTE".}

\item{...}{Placeholders.}
}
\value{
a \code{SpatRaster} object
}
\description{
The \code{process_edgar()} function imports and cleans EDGAR inventory
yearly VOC speciation data, returning a single \code{SpatRaster} object.
}
\note{
Layer names of the returned \code{SpatRaster} object contain the variable acronym,
and date.
}
\examples{
## NOTE: current version only supports processing of VOC speciation data
##       in .nc format
\dontrun{
edgar_voc <- process_edgar(
  path = "./data/edgar/voc/data_files/"
  year = c(2018, 2021),
  voc = 1:3,
  sector_voc = c(
    "AGRICULTURE",
    "BUILDINGS")
)
}
}
\author{
Mariana Kassien
}
