% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_geos}
\alias{process_geos}
\title{Import and clean GEOS-CF data downloaded with
\code{download_geos_cf_data} or \code{download_data(dataset_name = "geos")}. Function
returns a SpatRast object containing the user-defined variables of interest.
Layer names indicate the variable, pressure level, date (YYYYMMDD), and, if
applicable, the hour (HHMMSS).}
\usage{
process_geos(
  date_start = "2018-01-01",
  date_end = "2018-01-01",
  variable = NULL,
  path
)
}
\arguments{
\item{date_start}{character(1). length of 10. Format "YYYY-MM-DD".}

\item{date_end}{character(1). length of 10. Format "YYYY-MM-DD".}

\item{variable}{character(1). GEOS-CF variable name(s).}

\item{path}{character(1). Directory with downloaded GEOS-CF
netCDF files.}
}
\value{
a SpatRaster object;
}
\description{
Import and clean GEOS-CF data downloaded with
\code{download_geos_cf_data} or \code{download_data(dataset_name = "geos")}. Function
returns a SpatRast object containing the user-defined variables of interest.
Layer names indicate the variable, pressure level, date (YYYYMMDD), and, if
applicable, the hour (HHMMSS).
}
\author{
Mitchell Manware
}
