% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_geos}
\alias{process_geos}
\title{Process atmospheric composition data}
\usage{
process_geos(
  date = c("2018-01-01", "2018-01-01"),
  variable = NULL,
  path = NULL
)
}
\arguments{
\item{date}{character(2). length of 10. Format "YYYY-MM-DD".}

\item{variable}{character(1). GEOS-CF variable name(s).}

\item{path}{character(1). Directory with downloaded netCDF (.nc4) files.}
}
\value{
a SpatRaster object;
}
\description{
The \code{process_geos()} function imports and cleans raw atmospheric
composition data, returning a single SpatRaster object.
}
\note{
Layer names of the returned SpatRaster object contain the variable,
pressure level, date, and hour.
}
\author{
Mitchell Manware
}
