% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_geos_collection}
\alias{process_geos_collection}
\title{Identify GEOS-CF collection based on user-defined file paths}
\usage{
process_geos_collection(
  path,
  collection = FALSE,
  date = FALSE,
  datetime = FALSE
)
}
\arguments{
\item{path}{character(1). File path to GEOS-CF data file.}

\item{collection}{logical(1). Identifies and returns GEOS-CF collection
name(s) based on provided file path(s).}

\item{date}{logical(1). Identifies and returns date sequence (YYYYMMDD) based
on provided file path(s).}

\item{datetime}{logical(1). Identifies and returns date time sequence
(YYYYMoMoDDHHMiMi) based on provided file path(s).}
}
\value{
character
}
\description{
Identify GEOS-CF collection based on user-defined file paths
}
