% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_gmted}
\alias{process_gmted}
\title{Process elevation data}
\usage{
process_gmted(variable = NULL, path = NULL)
}
\arguments{
\item{variable}{vector(1). Vector containing the GMTED statistic first and
the resolution second. (Example: variable = c("Breakline Emphasis",
"7.5 arc-seconds")).}

\item{path}{character(1). Directory with downloaded GMTED  "*_grd"
folder containing .adf files.}
}
\value{
a SpatRaster object
}
\description{
The \code{process_gmted()} function imports and cleans raw elevation data,
returning a single SpatRaster object.
}
\note{
SpatRaster layer name indicates selected variable and resolution.
}
\author{
Mitchell Manware
}
