% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_gmted}
\alias{process_gmted}
\title{Import and clean Global Multi-resolution Terrain Elevation Data (GMTED2010)
downloaded with \code{download_gmted} or \code{download_data(dataset_name = "gmted")}.
Function returns a SpatRast object containing the user-defined variable
of interest at specified resolution. Layer name indicates variable and
resolution.}
\usage{
process_gmted(variable = NULL, path = NULL)
}
\arguments{
\item{variable}{vector(1). Vector containing the GMTED statistic first and
the resolution second. (Example: variable = c("Breakline Emphasis",
"7.5 arc-seconds")).}

\item{path}{character(1). Directory with downloaded GEOS-CF
the "*_grd" folder containing .adf files.}
}
\value{
a SpatRaster object
}
\description{
Import and clean Global Multi-resolution Terrain Elevation Data (GMTED2010)
downloaded with \code{download_gmted} or \code{download_data(dataset_name = "gmted")}.
Function returns a SpatRast object containing the user-defined variable
of interest at specified resolution. Layer name indicates variable and
resolution.
}
\author{
Mitchell Manware
}
