% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_gmted_codes}
\alias{process_gmted_codes}
\title{Process elevation statistic and resolution codes}
\usage{
process_gmted_codes(
  string,
  statistic = FALSE,
  resolution = FALSE,
  invert = FALSE
)
}
\arguments{
\item{string}{character(1). File path to GMTED data file.}

\item{statistic}{logical(1). Matches statistic to statistic code.}

\item{resolution}{logical(1). Matches resolution to resolution code.}

\item{invert}{logical(1). Default = FALSE. \code{invert = TRUE} assumes \code{string}
provides statistic or resolution code, and returns full length statistic
or resolution.}
}
\value{
character
}
\description{
Identify the GMTED statistic and resolution based on the file path. Convert
statistic and resolution to/from full string to/from statistic and
resolution code.
}
\keyword{auxiliary}
