% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_hms}
\alias{process_hms}
\title{Import and clean wildfire smoke plume coverage data from NOAA Hazard
Mapping System Fire and Smoke Product.}
\usage{
process_hms(
  date_start = "2018-01-01",
  date_end = "2018-01-01",
  variable = c("Light", "Medium", "Heavy"),
  path = "./input/noaa_hms/raw/"
)
}
\arguments{
\item{date_start}{character(1). length of 10. Start date of downloaded data.
Format YYYY-MM-DD (ex. September 1, 2023 = "2023-09-01").}

\item{date_end}{character(1). length of 10. End date of downloaded data.
Format YYYY-MM-DD (ex. September 10, 2023 = "2023-09-10").}

\item{variable}{character(1). "Light", "Medium", or "Heavy".}

\item{path}{character(1). Directory with downloaded NOAA HMS data files.}
}
\value{
a SpatVector object;
}
\description{
Import and clean wildfire smoke plume coverage data from NOAA Hazard
Mapping System Fire and Smoke Product.
}
\author{
Mitchell Manware.
}
