% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_hms}
\alias{process_hms}
\title{Import and clean wildfire smoke plume coverage data from NOAA Hazard
Mapping System Fire and Smoke Product.}
\usage{
process_hms(
  date = c("2018-01-01", "2018-01-01"),
  variable = c("Light", "Medium", "Heavy"),
  path = NULL
)
}
\arguments{
\item{date}{character(2). length of 10 each.
Start/end date of downloaded data.
Format YYYY-MM-DD (ex. September 1, 2023 = "2023-09-01").}

\item{variable}{character(1). "Light", "Medium", or "Heavy".}

\item{path}{character(1). Directory with downloaded NOAA HMS data files.}
}
\value{
a SpatVector object;
}
\description{
Import and clean wildfire smoke plume coverage data from NOAA Hazard
Mapping System Fire and Smoke Product.
}
\author{
Mitchell Manware.
}
