% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_hms}
\alias{process_hms}
\title{Process wildfire smoke data}
\usage{
process_hms(
  date = c("2018-01-01", "2018-01-01"),
  variable = c("Light", "Medium", "Heavy"),
  path = NULL
)
}
\arguments{
\item{date}{character(2). length of 10 each.
Start/end date of downloaded data.
Format YYYY-MM-DD (ex. September 1, 2023 = "2023-09-01").}

\item{variable}{character(1). "Light", "Medium", or "Heavy".}

\item{path}{character(1). Directory with downloaded NOAA HMS data files.}
}
\value{
a SpatVector or character object
}
\description{
The \code{process_hms()} function imports and cleans raw wildfire smoke
plume coverage data, returning a single SpatVector object.
}
\note{
\code{process_hms()} will return a character object if there are no wildfire
smoke plumes present for the selected dates and density. The returned
character will contain the selected density value and the sequence of dates
for which no wildfire smoke plumes were detected.
}
\examples{
process_hms(
  date = c("2018-12-30", "2019-01-01"),
  variable = "Light",
  path = "your/path/"
)
Light smoke plume polygons absent from 2018-12-30 to 2019-01-01. Returning vector of dates.
[1] "Light"    "20181230" "20181231" "20190101"
}
\author{
Mitchell Manware
}
