% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_huc}
\alias{process_huc}
\title{Retrieve Hydrologic Unit Code (HUC) data}
\usage{
process_huc(path, layer_name = NULL, huc_level = NULL, huc_header = NULL, ...)
}
\arguments{
\item{path}{character. Path to the file or the directory containing HUC data.}

\item{layer_name}{character(1). Layer name in the \code{path}}

\item{huc_level}{character(1). Field name of HUC level}

\item{huc_header}{character(1). The upper level HUC code header to extract
lower level HUCs.}

\item{...}{Arguments passed to \code{nhdplusTools::get_huc()}}
}
\value{
a \code{SpatVector} object
}
\description{
Retrieve Hydrologic Unit Code (HUC) data
}
\examples{

library(terra)
getf <- "../../../../group/set/Projects/PrestoGP_Pesticides/input/WBD-National/WBD_National_GDB.gdb"
terra::vector_layers(getf)
test <- process_huc(
  getf,
  layer_name = "WBDHU8",
  huc_level = "huc8"
)
test <- process_huc(
  "",
  layer_name = NULL,
  huc_level = NULL,
  huc_header = NULL,
  id = "020974",
  type = "huc06"
)
}
\seealso{
\code{\link[nhdplusTools:get_huc]{nhdplusTools::get_huc}}
}
\author{
Insang Song
}
