% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_koppen_geiger}
\alias{process_koppen_geiger}
\title{Process Koeppen-Geiger climate data}
\usage{
process_koppen_geiger(path, year = NULL)
}
\arguments{
\item{path}{character(1). Path to Koppen-Geiger
climate zone raster file}

\item{year}{data year. Not applicable for this function.}
}
\value{
SpatRaster object
}
\description{
Process Koeppen-Geiger climate data
}
\author{
Insang Song
}
