% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_locs_vector}
\alias{process_locs_vector}
\title{Import and prepare site point locations for covariate calculations.}
\usage{
process_locs_vector(locs, crs, radius)
}
\arguments{
\item{locs}{data.frame(1). Data frame containing columns for unique
identifier, latitute, and longitude. Latitude and longitude columns \strong{must}
be named "lat" and "lon", respectively.}

\item{crs}{Coordinate reference system (CRS) description utilizing
\code{terra::crs()}.}

\item{radius}{integer(1). Circular buffer size (meters).}
}
\value{
SpatVector
}
\description{
Import and prepare site point locations for covariate calculations.
}
