% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_modis_daily}
\alias{process_modis_daily}
\title{A single-date MODIS worker for parallelization}
\usage{
process_modis_daily(
  locs = NULL,
  from = NULL,
  locs_id = "site_id",
  date = NULL,
  name_extracted = NULL,
  fun_summary = "mean",
  radius = 0L
)
}
\arguments{
\item{locs}{SpatVector/sf/sftime object. Locations where MODIS values
are summarized.}

\item{from}{SpatRaster. Preprocessed objects.}

\item{locs_id}{character(1). Field name where unique site identifiers
are stored. Default is \code{"site_id"}}

\item{date}{Date(1). date to query.}

\item{name_extracted}{character. Names of calculated covariates.}

\item{fun_summary}{function. Summary function for
multilayer rasters. Passed to \code{foo}. See \link[exactextractr:exact_extract]{exactextractr::exact_extract}
for details.}

\item{radius}{numeric. Radius to generate circular buffers.}
}
\value{
A data.frame object.
}
\description{
modis_worker operates at six MODIS/VIIRS products
(MOD11A1, MOD13A2, MOD06_L2, VNP46A2, MOD09GA, and MCD19A2)
on a daily basis. Given that the raw hdf files are downloaded from
NASA, standard file names include a data retrieval date flag starting
with A. Leveraging that piece of information, the function will select
files of scope on the date of interest. Please note that this function
does not provide a function to filter swaths or tiles, so it is strongly
recommended to check and pre-filter the file names at users' discretion.
}
\author{
Insang Song
}
