% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates_support.R
\name{process_modis_sds}
\alias{process_modis_sds}
\title{Selected MODIS sinusoidal grid product subdataset name selector}
\usage{
process_modis_sds(
  product = c("MOD11A1", "MOD13A2", "MOD09GA", "MCD19A2"),
  custom_sel = NULL
)
}
\arguments{
\item{product}{character(1). Product code.}

\item{custom_sel}{character(1). Custom filter.
If this value is not NULL, preset filter is
overridden.}
}
\value{
A character object that conforms to the regular
expression. Details of regular expression in R can be found in \link{regexp}.
}
\description{
Four presets are supported. \code{custom_sel} supersedes
presets of \code{product} values.
}
\seealso{
\link{calc_modis}
}
\author{
Insang Song
}
