% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw.R
\name{process_modis_swath}
\alias{process_modis_swath}
\title{Mosaic MODIS swaths}
\usage{
process_modis_swath(
  paths,
  date_in,
  get_var = c("Cloud_Fraction_Day", "Cloud_Fraction_Night"),
  suffix = ":mod06:",
  resolution = 0.025,
  ...
)
}
\arguments{
\item{paths}{character. Full paths of hdf files.}

\item{date_in}{character(1). Date to query.}

\item{get_var}{character. One of \code{"Cloud_Fraction_Day"} or
\code{"Cloud_Fraction_Night"} (which are available in MOD06_L2)}

\item{suffix}{character(1). Should be formatted \verb{:\{product\}:},
e.g., \verb{:mod06:}}

\item{resolution}{numeric(1). Resolution of output raster.
Unit is degree.}

\item{...}{For internal use.}
}
\value{
SpatRaster object. CRS is \code{"EPSG:4326"}.
}
\description{
This function will return a SpatRaster object with
values of selected subdatasets. Swath data include curvilinear
grids, which require warping/rectifying the original curvilinear grids
into rectilinear grids. The function internally warps each of inputs
then mosaic the warped images into one large SpatRaster object.
}
\seealso{
\itemize{
\item \link{process_modis_warp}
\item \href{https://gdal.org/drivers/raster/hdf4.html}{GDAL HDF4 driver documentation}
\item \link[terra:describe]{terra::describe}: to list the full subdataset list with \code{sds = TRUE}
}
}
\author{
Insang Song
}
