% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_narr}
\alias{process_narr}
\title{Process meteorological data}
\usage{
process_narr(
  date = c("2023-09-01", "2023-09-01"),
  variable = NULL,
  path = NULL
)
}
\arguments{
\item{date}{character(2). length of 10 each. Format "YYYY-MM-DD".}

\item{variable}{character(1). Variable name acronym.}

\item{path}{character(1). Directory with downloaded netCDF (.nc) files.}
}
\value{
a \code{SpatRaster} object
}
\description{
The \code{process_narr()} function imports and cleans raw meteorological
data, returning a single \code{SpatRaster} object.
}
\note{
Layer names of the returned \code{SpatRaster} object contain the variable acronym,
pressure level, and date.
}
\author{
Mitchell Manware
}
