% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw.R
\name{process_nei}
\alias{process_nei}
\title{Process raw National Emission Inventory (NEI) data}
\usage{
process_nei(path = "./input/nei/", year = 2017, auxfile)
}
\arguments{
\item{path}{character(1). Path to the directory with NEI CSV files}

\item{year}{integer(1). Data year.
Currently only accepts \code{c(2017, 2020)}}

\item{auxfile}{character(1) or SpatVector.
Path to or SpatVector object of the county boundaries in \code{year}}
}
\value{
SpatVector object. The third column stores the total emission.
}
\description{
NEI data comprises multiple csv files where emissions of
50+ pollutants are recorded at county level in ten EPA regions.
This function will return a combined table of NEI data.
}
\author{
Insang Song, Ranadeep Daw
}
