% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_nlcd}
\alias{process_nlcd}
\title{Process land cover data}
\usage{
process_nlcd(path = NULL, year = 2021)
}
\arguments{
\item{path}{character giving nlcd data path}

\item{year}{numeric giving the year of NLCD data used}
}
\value{
a SpatRaster object
}
\description{
The \code{process_nlcd()} function imports and cleans raw land cover data,
returning a single SpatRaster object.

Reads NLCD file of selected \code{year}.
}
\author{
Eva Marques, Insang Song
}
