% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw.R
\name{process_nlcd}
\alias{process_nlcd}
\title{Process raw National Landuse Classification Dataset}
\usage{
process_nlcd(path, year = 2021)
}
\arguments{
\item{path}{character giving nlcd data path}

\item{year}{numeric giving the year of NLCD data used}
}
\value{
SpatRaster object.
}
\description{
Reads NLCD file of selected \code{year}.
}
\author{
Eva Marques
}
