% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_olm}
\alias{process_olm}
\title{Process OpenLandMap data}
\usage{
process_olm(path = NULL, ...)
}
\arguments{
\item{path}{character giving OpenLandMap data path}

\item{...}{Placeholders.}
}
\value{
SpatRaster
}
\description{
Process OpenLandMap data
}
\author{
Insang Song
}
