% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_sedac_groads}
\alias{process_sedac_groads}
\title{Import and clean groads data from NASA Socioeconomic Data and Applications Center (SEDAC).}
\usage{
process_sedac_groads(path = NULL)
}
\arguments{
\item{path}{character(1). Path to geodatabase or shapefiles.}
}
\value{
a SpatVector.
}
\description{
Import and clean groads data from NASA Socioeconomic Data and Applications Center (SEDAC).
}
\note{
U.S. context.
}
\author{
Insang Song
}
