% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_sedac_groads}
\alias{process_sedac_groads}
\title{Process roads data}
\usage{
process_sedac_groads(path = NULL, ...)
}
\arguments{
\item{path}{character(1). Path to geodatabase or shapefiles.}

\item{...}{Placeholders.}
}
\value{
a \code{SpatVector} object
}
\description{
The \code{process_sedac_groads()} function imports and cleans raw road data,
returning a single \code{SpatVector} object.
}
\note{
U.S. context. The returned \code{SpatVector} object contains a
\verb{$description} column to represent the temporal range covered by the
dataset. For more information, see \url{https://sedac.ciesin.columbia.edu/data/set/groads-global-roads-open-access-v1/metadata}.
}
\author{
Insang Song
}
