% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_sedac_population}
\alias{process_sedac_population}
\title{Process population density data}
\usage{
process_sedac_population(path = NULL, ...)
}
\arguments{
\item{path}{character(1). Path to GeoTIFF (.tif) or netCDF (.nc) file.}

\item{...}{Placeholders.}
}
\value{
a \code{SpatRaster} object
}
\description{
The \code{process_secac_population()} function imports and cleans raw
population density data, returning a single \code{SpatRaster} object.
}
\author{
Mitchell Manware
}
