% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R, R/process_raw.R
\name{process_tri}
\alias{process_tri}
\title{Prepare Toxic Release Inventory (TRI) data}
\usage{
process_tri(path = "./input/tri/", year = 2018L)

process_tri(path = "./input/tri/", year = 2018L)
}
\arguments{
\item{path}{character(1). Path to the directory with TRI CSV files}

\item{variables}{integer. Column index of TRI data.}

\item{domain_year}{integer. Year to select.
Default is 2018.}
}
\value{
SpatVector (points) object.

SpatVector object of TRI locations in the selected `year``
}
\description{
Prepare Toxic Release Inventory (TRI) data

Process raw TRI data
}
\note{
TRI data is available in USEPA.

U.S. context.
}
\references{
https://www.epa.gov/toxics-release-inventory-tri-program/tri-data-and-tools
}
\author{
Insang Song, Mariana Kassien
}
