% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_tri}
\alias{process_tri}
\title{Prepare Toxic Release Inventory (TRI) data}
\usage{
process_tri(
  path = "./input/tri/",
  year = 2018,
  variables = c(1, 13, 12, 14, 20, 34, 36, 47, 48, 49),
  ...
)
}
\arguments{
\item{path}{character(1). Path to the directory with TRI CSV files}

\item{year}{integer(1). Single year to select.}

\item{variables}{integer. Column index of TRI data.}

\item{...}{Placeholders.}
}
\value{
SpatVector (points) object in \code{year}.
}
\description{
Prepare Toxic Release Inventory (TRI) data
}
\note{
TRI data is available in USEPA. Visit the page to see the available
year and column descriptions.
}
\references{
https://www.epa.gov/toxics-release-inventory-tri-program/tri-data-and-tools
}
\author{
Insang Song, Mariana Kassien
}
