% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw.R
\name{process_tri}
\alias{process_tri}
\title{Process raw TRI data}
\usage{
process_tri(path = "./input/tri/", year = 2018L)
}
\arguments{
\item{path}{character(1). Path to the directory with TRI CSV files}

\item{domain_year}{integer. Year to select.
Default is 2018.}
}
\value{
SpatVector object of TRI locations in the selected `year``
}
\description{
Process raw TRI data
}
\note{
U.S. context.
}
\author{
Insang Song, Mariana Kassien
}
