% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{rename_time}
\alias{rename_time}
\title{Rename the time column of a sftime object}
\usage{
rename_time(x, newname)
}
\arguments{
\item{x}{a sftime object}

\item{newname}{character for new time column name}
}
\value{
a sftime object
}
\description{
Rename the time column of a sftime object
}
\author{
Eva Marques
}
