% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{sf_as_mysftime}
\alias{sf_as_mysftime}
\title{Convert a sf object to mysftime}
\usage{
sf_as_mysftime(x, timename)
}
\arguments{
\item{x}{a sf}

\item{timename}{character: name of time column in x}
}
\value{
a sftime object
}
\description{
Convert a sf object to mysftime
}
\author{
Eva Marques
}
\keyword{spacetime}
