% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{sftime_as_mysftime}
\alias{sftime_as_mysftime}
\title{Convert a sftime object to mysftime}
\usage{
sftime_as_mysftime(x, timename)
}
\arguments{
\item{x}{a sftime}

\item{timename}{character: name of time column in x}
}
\value{
a sftime object with specific format (see check_mysftime() function)
}
\description{
Convert a sftime object to mysftime
}
\author{
Eva Marques
}
