% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{sftime_as_sf}
\alias{sftime_as_sf}
\title{Simplify an sftime to sf class}
\usage{
sftime_as_sf(x, keeptime = TRUE)
}
\arguments{
\item{x}{a sftime}

\item{keeptime}{boolean: TRUE if user wants to keep time column
as simple column (default = TRUE)}
}
\value{
a sf object
}
\description{
Simplify an sftime to sf class
}
\author{
Eva Marques
}
