% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{sftime_as_spatvector}
\alias{sftime_as_spatvector}
\title{Convert an \code{sftime} to a \code{SpatVector}}
\usage{
sftime_as_spatvector(x)
}
\arguments{
\item{x}{an \code{sftime} object}
}
\value{
a \code{SpatVector} object
}
\description{
Convert an \code{sftime} object to a \code{SpatVector} object.
}
\seealso{
\link[terra:vect]{terra::vect}
}
\author{
Eva Marques
}
\keyword{spacetime}
