% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{spatraster_as_sftime}
\alias{spatraster_as_sftime}
\title{Create a sftime from a terra::SpatRaster}
\usage{
spatraster_as_sftime(x, varname, timename = "time")
}
\arguments{
\item{x}{a terra::SpatRaster}

\item{varname}{character for variable column name in the sftime}

\item{timename}{character for time column name in the sftime
(default: "time")}
}
\value{
a sftime object
}
\description{
Create a sftime from a terra::SpatRaster
}
\author{
Eva Marques
}
\keyword{spacetime}
