% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{spatrds_as_sftime}
\alias{spatrds_as_sftime}
\title{Create a sftime from a terra::SpatRasterDataset}
\usage{
spatrds_as_sftime(x, timename = "time")
}
\arguments{
\item{x}{a terra::SpatRasterDataset (~ list of named SpatRasters)}

\item{timename}{character for time column name in the sftime
(default: "time")}
}
\value{
a sftime object
}
\description{
Create a sftime from a terra::SpatRasterDataset
}
\author{
Eva Marques
}
