% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{spatvector_as_sftime}
\alias{spatvector_as_sftime}
\title{Create a sftime from a terra::SpatVector}
\usage{
spatvector_as_sftime(x, timename = "time")
}
\arguments{
\item{x}{a terra::SpatVector}

\item{timename}{character for time column name in x
(default: "time")}
}
\value{
a sftime object
}
\description{
Create a sftime from a terra::SpatVector
}
\author{
Eva Marques
}
