% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_test_support.R
\name{test_download_functions}
\alias{test_download_functions}
\title{Apply download function-specific unit tests}
\usage{
test_download_functions(
  directory_to_download = NULL,
  directory_to_save = directory_to_save,
  commands_path = commands_path,
  url_status = url_status
)
}
\arguments{
\item{directory_to_download}{directory to test download}

\item{directory_to_save}{directory to test saving}

\item{commands_path}{file path with download commands}

\item{url_status}{logical vector for URL status = 200}
}
\description{
Apply download function-specific unit tests
}
